[IMP] point_of_sale: avoid running useless mobile test
Before this commit all tests were run on both desktop and mobile, even
if the test was only relevant for one of the two platforms.

Mobile test should only test the UI of the mobile version of the PoS
to be able to check that all the features are working properly on mobile

Desktop version is the main version of the test, which will test
business logic and features.
