+++ addons/point_of_sale/tests/__init__.py
@@ -26,3 +26,4 @@ from . import test_res_config_settings
 from . import test_pos_product_variants
 from . import test_generic_localization
 from . import test_order_receipt
+from . import test_frontend_desktop_only
