+++ addons/point_of_sale/tests/test_frontend.py
@@ -1,32 +1,20 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 import inspect
-import logging
-import io
-
-from PIL import Image
 from contextlib import contextmanager
-from unittest.mock import patch
-from unittest import skip
-from odoo import Command, api
-
-from odoo.tools import BinaryBytes, DEFAULT_SERVER_DATE_FORMAT
-from odoo.tests import tagged, loaded_demo_data
-from odoo.addons.account.tests.common import TestTaxCommon, AccountTestInvoicingHttpCommon
-from odoo.addons.point_of_sale.tests.common_setup_methods import setup_product_combo_items
 from datetime import date, timedelta
+from unittest import skip
+from unittest.mock import patch
+
+from odoo import Command
+from odoo.tests import tagged
+from odoo.tools import DEFAULT_SERVER_DATE_FORMAT
+
+from odoo.addons.account.tests.common import (
+    AccountTestInvoicingHttpCommon,
+    TestTaxCommon,
+)
 from odoo.addons.point_of_sale.tests.common import archive_products
-from odoo.exceptions import UserError
-from freezegun import freeze_time
-
-_logger = logging.getLogger(__name__)
-
-
-def _create_image(color: int | str = 0, dims=(1920, 1080), format='JPEG'):
-    f = io.BytesIO()
-    Image.new('RGB', dims, color).save(f, format)
-    f.seek(0)
-    return BinaryBytes(f.read())
 
 
 class TestPointOfSaleHttpCommon(AccountTestInvoicingHttpCommon):
@@ -645,12 +633,6 @@ class TestUi(TestPointOfSaleHttpCommon):
     def test_01_pos_basic_order(self):
         self.start_pos_tour('pos_pricelist')
 
-    def test_pos_basic_order_02_decimal_order_quantity(self):
-        self.start_pos_tour('pos_basic_order_02_decimal_order_quantity')
-
-    def test_pos_basic_order_03_tax_position(self):
-        self.start_pos_tour('pos_basic_order_03_tax_position')
-
     def test_floating_order_tour(self):
         self.start_pos_tour('FloatingOrderTour')
 
@@ -688,26 +670,6 @@ class TestUi(TestPointOfSaleHttpCommon):
             message = self.env['mail.message'].search([('model', '=', 'pos.order'), ('res_id', '=', order.id)], limit=1)
             self.assertEqual(len(message.attachment_ids), 2, "Should have 2 attachments when basic receipt is True")
 
-    @skip('Temporary to fast merge new valuation')
-    def test_02_pos_with_invoiced(self):
-        self.pos_user.write({
-            'group_ids': [
-                (4, self.env.ref('account.group_account_invoice').id),
-            ]
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'ChromeTour', login="pos_user")
-        n_invoiced = self.env['pos.order'].search_count([('account_move', '!=', False)])
-        n_paid = self.env['pos.order'].search_count([('state', '=', 'paid')])
-        self.assertEqual(n_invoiced, 1, 'There should be 1 invoiced order.')
-        self.assertEqual(n_paid, 2, 'There should be 2 paid order.')
-        last_order = self.env['pos.order'].search([], limit=1, order="id desc")
-        self.assertEqual(last_order.lines[0].price_subtotal, 30.0)
-        self.assertEqual(last_order.lines[0].price_subtotal_incl, 30.0)
-        # Check if session name contains config name as prefix
-        self.assertEqual(self.main_pos_config.name in last_order.session_id.name, True)
-
     def test_04_product_configurator(self):
         # Making one attribute inactive to verify that it doesn't show
         configurable_product = self.env['product.product'].search([('name', '=', 'Configurable Chair'), ('available_in_pos', '=', True)], limit=1)
@@ -715,29 +677,6 @@ class TestUi(TestPointOfSaleHttpCommon):
         fabrics_line.product_template_value_ids[1].ptav_active = False
         self.main_pos_config.with_user(self.pos_user).open_ui()
         self.start_pos_tour('ProductConfiguratorTour')
-
-    def test_optional_product(self):
-        # optional product in pos
-        image = _create_image(color="orange")
-
-        self.small_shelf.write({'image_1920': image})
-
-        self.desk_pad.write({'pos_optional_product_ids': [
-            Command.set([ self.small_shelf.id ])
-        ]})
-
-        self.letter_tray.write({
-            'pos_optional_product_ids': [Command.set([self.configurable_chair.id])],
-            'barcode': 'lettertray'
-        })
-        # Case 1: Images ON → images must be visible in optional product dialog
-        self.main_pos_config.show_product_images = True
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_optional_product')
-
-        # Case 2: Images OFF → product images should not be shown
-        self.main_pos_config.show_product_images = False
-        self.start_pos_tour('test_optional_product_image_not_display')
 
     @skip('Temporary to fast merge new valuation')
     def test_05_ticket_screen(self):
@@ -750,17 +689,6 @@ class TestUi(TestPointOfSaleHttpCommon):
         self.main_pos_config.with_user(self.pos_user).open_ui()
         self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'TicketScreenTour', login="pos_user")
         self.env['res.lang']._lang_get(self.pos_user.lang).write({'date_format': 'MM/dd/yyyy', 'time_format': 'HH:mm:ss'})
-
-    def test_06_tip_screen(self):
-        self.main_pos_config.write({'set_tip_after_payment': True, 'iface_tipproduct': True, 'tip_product_id': self.env.ref('point_of_sale.product_product_tip')})
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'PosTipAfterPaymentTour', login="pos_user")
-
-        orders = self.env['pos.order'].search([], limit=11, order="id desc")
-        order_tips = [o.tip_amount for o in orders]
-
-        order_tips.sort()
-        self.assertEqual(order_tips, [0.0, 0.0, 0.0, 0.0, 0.0, 0.3, 0.8, 1.0, 1.5, 2.0, 10.0])
 
     def test_product_information_screen_admin(self):
         '''Consider this test method to contain a test tour with miscellaneous tests/checks that require admin access.
@@ -776,1739 +704,11 @@ class TestUi(TestPointOfSaleHttpCommon):
         self.main_pos_config.with_user(self.pos_admin).open_ui()
         self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'CheckProductInformation', login="pos_admin")
 
-    def test_fixed_tax_negative_qty(self):
-        """ Assert the negative amount of a negative-quantity orderline
-            with zero-amount product with fixed tax.
-        """
-
-        # setup the zero-amount product
-        tax_received_account = self.env['account.account'].create({
-            'name': 'TAX_BASE',
-            'code': 'TBASE',
-            'account_type': 'asset_current',
-        })
-        fixed_tax = self.env['account.tax'].create({
-            'name': 'fixed amount tax',
-            'amount_type': 'fixed',
-            'amount': 1,
-            'invoice_repartition_line_ids': [
-                (0, 0, {'repartition_type': 'base'}),
-                (0, 0, {
-                    'repartition_type': 'tax',
-                    'account_id': tax_received_account.id,
-                }),
-            ],
-            'price_include_override': 'tax_excluded',
-        })
-        zero_amount_product = self.env['product.product'].create({
-            'name': 'Zero Amount Product',
-            'available_in_pos': True,
-            'list_price': 0,
-            'taxes_id': [(6, 0, [fixed_tax.id])],
-            'categ_id': self.env.ref('product.product_category_services').id,
-        })
-
-        # Make an order with the zero-amount product from the frontend.
-        # We need to do this because of the fix in the "compute_all" port.
-        self.main_pos_config.write({'iface_tax_included': 'total'})
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'FixedTaxNegativeQty', login="pos_user")
-        pos_session = self.main_pos_config.current_session_id
-
-        # Close the session and check the session journal entry.
-        pos_session.action_pos_session_validate()
-
-        lines = pos_session.move_id.line_ids.sorted('balance')
-
-        # order in the tour is paid using the bank payment method.
-        bank_pm = self.main_pos_config.payment_method_ids.filtered(lambda pm: pm.name == 'Bank')
-
-        self.assertEqual(lines[0].account_id, bank_pm.receivable_account_id or self.env.company.account_default_pos_receivable_account_id)
-        self.assertAlmostEqual(lines[0].balance, -1)
-        self.assertEqual(lines[1].account_id, self.env.company.income_account_id)
-        self.assertAlmostEqual(lines[1].balance, 0)
-        self.assertEqual(lines[2].account_id, tax_received_account)
-        self.assertAlmostEqual(lines[2].balance, 1)
-
-    def test_change_without_cash_method(self):
-        #create bank payment method
-        bank_pm = self.env['pos.payment.method'].create({
-            'name': 'Bank',
-            'receivable_account_id': self.env.company.account_default_pos_receivable_account_id.id,
-            'is_cash_count': False,
-            'split_transactions': False,
-            'company_id': self.env.company.id,
-        })
-        self.main_pos_config.write({'payment_method_ids': [(6, 0, bank_pm.ids)]})
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'PaymentScreenTour2', login="pos_user")
-
-    def test_rounding_up(self):
-        rouding_method = self.env['account.cash.rounding'].create({
-            'name': 'Rounding up',
-            'rounding': 0.05,
-            'rounding_method': 'UP',
-        })
-
-        self.env['product.product'].create({
-            'name': 'Product Test',
-            'available_in_pos': True,
-            'list_price': 1.96,
-            'taxes_id': False,
-        })
-
-        self.main_pos_config.write({
-            'rounding_method': rouding_method.id,
-            'cash_rounding': True,
-            'only_round_cash_method': True,
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'PaymentScreenRoundingUp', login="pos_user")
-
-    def test_rounding_down(self):
-        rouding_method = self.env['account.cash.rounding'].create({
-            'name': 'Rounding down',
-            'rounding': 0.05,
-            'rounding_method': 'DOWN',
-        })
-
-        self.env['product.product'].create({
-            'name': 'Product Test',
-            'available_in_pos': True,
-            'list_price': 1.98,
-            'taxes_id': False,
-        })
-
-        self.main_pos_config.write({
-            'rounding_method': rouding_method.id,
-            'cash_rounding': True,
-            'only_round_cash_method': True,
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'PaymentScreenRoundingDown', login="pos_user")
-        self.env["pos.order"].search([('state', '=', 'draft')]).write({'state': 'cancel'})
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'PaymentScreenTotalDueWithOverPayment', login="pos_user")
-
-    def test_pos_closing_cash_details(self):
-        """Test cash difference *loss* at closing.
-        """
-        self.main_pos_config.open_ui()
-        current_session = self.main_pos_config.current_session_id
-        current_session.post_closing_cash_details(0)
-        current_session.close_session_from_ui()
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'CashClosingDetails', login="pos_user")
-        self.assertEqual(self.main_pos_config.last_session_closing_cash, 50.0)
-        cash_diff_line = self.env['account.bank.statement.line'].search([
-            ('payment_ref', 'ilike', 'Cash difference observed during the counting (Loss)')
-        ])
-        self.assertAlmostEqual(cash_diff_line.amount, -1.00)
-
-    def test_cash_payments_should_reflect_on_next_opening(self):
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'OrderPaidInCash', login="pos_user")
-        self.assertEqual(self.main_pos_config.last_session_closing_cash, 25.0)
-
-    def test_pos_session_statistics_display(self):
-        """Test that POS session statistics are properly displayed in the UI."""
-        # For testing `opening_cash` and `paid_orders` in dashboard
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'SessionStatisticsDisplay', login="pos_user")
-
-        # For testing `draft_orders`
-        self.env['pos.order'].create({
-            'config_id': self.main_pos_config.id,
-            'session_id': self.main_pos_config.current_session_id.id,
-            'company_id': self.main_pos_config.company_id.id,
-            'amount_total': 10.0,
-            'amount_paid': 10.0,
-            'amount_tax': 0.0,
-            'amount_return': 0.0,
-            'to_invoice': False,
-            'partner_id': False,
-            'pricelist_id': self.main_pos_config.pricelist_id.id,
-            'pos_reference': '1000-004-00001',
-            'name': 'Order 1001',
-            'state': 'draft',
-            'lines': [(0, 0, {
-                'product_id': self.desk_pad.product_variant_id.id,
-                'price_unit': 10.00,
-                'discount': 0,
-                'qty': 1,
-                'tax_ids': False,
-                'price_subtotal': 10.00,
-                'price_subtotal_incl': 10.00,
-            })],
-        })
-
-        dashboard_statistics = self.main_pos_config.statistics_for_current_session
-
-        self.assertTrue(dashboard_statistics['date']['is_started'])
-        self.assertEqual(dashboard_statistics['cash']['raw_opening_cash'], 100.0)
-        self.assertEqual(dashboard_statistics['orders']['paid']['amount'], 45.0)
-        self.assertEqual(dashboard_statistics['orders']['paid']['count'], 2)
-        self.assertEqual(dashboard_statistics['orders']['draft']['amount'], 10.0)
-        self.assertEqual(dashboard_statistics['orders']['draft']['count'], 1)
-
-    def test_tax_control_button_visiblity(self):
-        self.main_pos_config.write({
-            'tax_regime_selection': False,
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_tax_control_button_visiblity')
-
-    def test_fiscal_position_no_tax(self):
-        #create a tax of 15% with price included
-        tax = self.env['account.tax'].create({
-            'name': 'Tax 15%',
-            'amount': 15,
-            'price_include_override': 'tax_included',
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-        })
-
-        #create a product with the tax
-        self.product = self.env['product.product'].create({
-            'name': 'Test Product',
-            'taxes_id': [(6, 0, [tax.id])],
-            'list_price': 100,
-            'available_in_pos': True,
-        })
-
-        #create a fiscal position that map the tax to no tax
-        fiscal_position = self.env['account.fiscal.position'].create({
-            'name': 'No Tax',
-        })
-
-        pricelist = self.env['product.pricelist'].create({
-            'name': 'Test Pricelist',
-        })
-
-        self.main_pos_config.write({
-            'tax_regime_selection': True,
-            'fiscal_position_ids': [(6, 0, [fiscal_position.id])],
-            'available_pricelist_ids': [(6, 0, [pricelist.id])],
-            'pricelist_id': pricelist.id,
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'FiscalPositionNoTax', login="pos_user")
-
-    def test_fiscal_position_inclusive_and_exclusive_tax(self):
-        """ Test the mapping of fiscal position for both Tax Inclusive ans Tax Exclusive"""
-        # create a fiscal position that map the tax
-        fiscal_position_1 = self.env['account.fiscal.position'].create({
-            'name': 'Incl. to Incl.',
-        })
-        fiscal_position_2 = self.env['account.fiscal.position'].create({
-            'name': 'Incl. to Excl.',
-        })
-        fiscal_position_3 = self.env['account.fiscal.position'].create({
-            'name': 'Excl. to Excl.',
-        })
-        fiscal_position_4 = self.env['account.fiscal.position'].create({
-            'name': 'Excl. to Incl.',
-        })
-
-        # create a tax with price included
-        tax_inclusive_1 = self.env['account.tax'].create({
-            'name': 'Tax incl.20%',
-            'amount': 20,
-            'price_include_override': 'tax_included',
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-        })
-        tax_exclusive_1 = self.env['account.tax'].create({
-            'name': 'Tax excl.20%',
-            'amount': 20,
-            'price_include_override': 'tax_excluded',
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-        })
-        self.env['account.tax'].create({
-            'name': 'Tax incl.10%',
-            'amount': 10,
-            'price_include_override': 'tax_included',
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-            'fiscal_position_ids': [Command.set((fiscal_position_1 | fiscal_position_4).ids)],
-            'original_tax_ids': [Command.set((tax_inclusive_1 | tax_exclusive_1).ids)],
-        })
-        self.env['account.tax'].create({
-            'name': 'Tax excl.10%',
-            'amount': 10,
-            'price_include_override': 'tax_excluded',
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-            'fiscal_position_ids': [Command.set((fiscal_position_2 | fiscal_position_3).ids)],
-            'original_tax_ids': [Command.set((tax_inclusive_1 | tax_exclusive_1).ids)],
-        })
-        self.test_product_1 = self.env['product.product'].create({
-            'name': 'Test Product 1',
-            'available_in_pos': True,
-            'list_price': 100,
-            'taxes_id': [(6, 0, [tax_inclusive_1.id])],
-        })
-
-        self.test_product_2 = self.env['product.product'].create({
-            'name': 'Test Product 2',
-            'available_in_pos': True,
-            'list_price': 100,
-            'taxes_id': [(6, 0, [tax_exclusive_1.id])],
-        })
-
-        # add the fiscal position to the PoS
-        self.main_pos_config.write({
-            'tax_regime_selection': True,
-            'fiscal_position_ids': [(6, 0, [
-                    fiscal_position_1.id,
-                    fiscal_position_2.id,
-                    fiscal_position_3.id,
-                    fiscal_position_4.id,
-                ])],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'FiscalPositionIncl', login="pos_user")
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'FiscalPositionExcl', login="pos_user")
-
-    def test_06_pos_discount_display_with_multiple_pricelist(self):
-        """ Test the discount display on the POS screen when multiple pricelists are used."""
-        test_product = self.env['product.template'].create({
-            'name': 'Test Product',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-        })
-
-        base_pricelist = self.env['product.pricelist'].create({
-            'name': 'base_pricelist',
-        })
-
-        self.env['product.pricelist.item'].create({
-            'pricelist_id': base_pricelist.id,
-            'product_tmpl_id': test_product.id,
-            'compute_price': 'percentage',
-            'applied_on': '1_product',
-            'percent_price': 30,
-        })
-
-        special_pricelist = self.env['product.pricelist'].create({
-            'name': 'special_pricelist',
-        })
-        self.env['product.pricelist.item'].create({
-            'pricelist_id': special_pricelist.id,
-            'base': 'pricelist',
-            'base_pricelist_id': base_pricelist.id,
-            'compute_price': 'percentage',
-            'applied_on': '3_global',
-            'percent_price': 10,
-        })
-
-        self.main_pos_config.write({
-            'pricelist_id': base_pricelist.id,
-            'available_pricelist_ids': [(6, 0, [base_pricelist.id, special_pricelist.id])],
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'FeedbackScreenDiscountWithPricelistTour', login="pos_user")
-
-    def test_07_product_combo(self):
-        self.env['decimal.precision'].search([('name', '=', 'Product Price')]).digits = 4
-        setup_product_combo_items(self)
-        self.desk_accessories_combo.sequence = 100
-        combo_product_sofa = self.env["product.template"].create(
-            {
-                "name": "Combo product Sofa",
-                "is_storable": True,
-                "available_in_pos": True,
-                "list_price": 40,
-            }
-        )
-        sofa_size_attribute = self.env['product.attribute'].create({
-            'name': 'Size',
-            'display_type': 'radio',
-            'create_variant': 'always',
-        })
-        sofa_color_attribute = self.env['product.attribute'].create({
-            'name': 'Color',
-            'display_type': 'radio',
-            'create_variant': 'always',
-        })
-        sofa_size_L = self.env['product.attribute.value'].create({
-            'name': 'L',
-            'attribute_id': sofa_size_attribute.id,
-        })
-        sofa_size_M = self.env['product.attribute.value'].create({
-            'name': 'M',
-            'attribute_id': sofa_size_attribute.id,
-        })
-        sofa_color_red = self.env['product.attribute.value'].create({
-            'name': 'red',
-            'attribute_id': sofa_color_attribute.id,
-        })
-        sofa_color_blue = self.env['product.attribute.value'].create({
-            'name': 'blue',
-            'attribute_id': sofa_color_attribute.id,
-        })
-
-        product_attribute_size = self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': combo_product_sofa.id,
-            'attribute_id': sofa_size_attribute.id,
-            'value_ids': [Command.set([sofa_size_M.id, sofa_size_L.id])],
-
-        })
-        self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': combo_product_sofa.id,
-            'attribute_id': sofa_color_attribute.id,
-            'value_ids': [Command.set([sofa_color_red.id, sofa_color_blue.id])],
-
-        })
-        product_attribute_size.product_template_value_ids[0].price_extra = 50
-        product_attribute_size.product_template_value_ids[1].price_extra = 100
-        self.sofa_combo = self.env["product.combo"].create(
-            {
-                "name": "Chairs Combo",
-                "combo_item_ids": [
-                    Command.create({
-                        "product_id": combo_product_sofa.product_variant_ids[0].id,
-                        "extra_price": 5,
-                    }),
-                    Command.create({
-                        "product_id": combo_product_sofa.product_variant_ids[1].id,
-                        "extra_price": 10,
-                    }),
-                ],
-            },
-        )
-        self.sofa_combo = self.env["product.product"].create(
-            {
-                "available_in_pos": True,
-                "list_price": 20,
-                "name": "Sofa Combo",
-                "type": "combo",
-                "uom_id": self.env.ref("uom.product_uom_unit").id,
-                "combo_ids": [
-                    Command.set([self.sofa_combo.id]),
-                ],
-            },
-        )
-        self.office_combo.write({
-            'lst_price': 50,
-            'barcode': 'SuperCombo',
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('ProductComboPriceTaxIncludedTour')
-        order = self.env['pos.order'].search([])
-        self.assertEqual(len(order.lines), 4, "There should be 4 order lines - 1 combo parent and 3 combo lines")
-        # check that the combo lines are correctly linked to each other
-        parent_line_id = self.env['pos.order.line'].search([('product_id.name', '=', 'Office Combo'), ('order_id', '=', order.id)])
-        combo_line_ids = self.env['pos.order.line'].search([('product_id.name', '!=', 'Office Combo'), ('order_id', '=', order.id)])
-        self.assertEqual(parent_line_id.combo_line_ids, combo_line_ids, "The combo parent should have 3 combo lines")
-        self.assertEqual(order.lines[1].price_unit, 18.67)
-        self.assertEqual(order.lines[2].price_unit, 30.00)
-        self.assertAlmostEqual(order.lines[3].price_unit, 10.33)
-        # In the future we might want to test also if:
-        #   - the combo lines are correctly stored in and restored from local storage
-        #   - the combo lines are correctly shared between the pos configs ( in cross ordering )
-
-    def test_07_product_combo_max_free_qty(self):
-        """ Test the max free quantity of a product combo."""
-        setup_product_combo_items(self)
-        self.office_combo.combo_ids[0].write({
-            'qty_free': 2,
-            'qty_max': 2,
-        })
-        self.office_combo.combo_ids[1].write({
-            'qty_free': 2,
-            'qty_max': 5,
-        })
-        self.combo_product_2.active = False
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('ProductComboMaxFreeQtyTour')
-
-    def test_line_configurators(self):
-        setup_product_combo_items(self)
-        self.env['product.combo.item'].create({
-            'combo_id': self.desks_combo.id,
-            'product_id': self.configurable_chair.product_variant_id.id,
-            'extra_price': 0,
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_line_configurators_product')
-        self.start_pos_tour('test_line_configurators_combo')
-
-    def test_07_pos_barcodes_scan(self):
-        barcode_rule = self.env.ref("point_of_sale.barcode_rule_client")
-        barcode_rule.pattern = barcode_rule.pattern + "|234"
-        # should in theory be changed in the JS code to `|^234`
-        # If not, it will fail as it will mistakenly match with the product barcode "0123456789"
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'BarcodeScanningTour', login="pos_user")
-
-    def test_08_show_tax_excluded(self):
-        # define a tax included tax record
-        tax = self.env['account.tax'].create({
-            'name': 'Tax 10% Included',
-            'amount_type': 'percent',
-            'amount': 10,
-            'price_include_override': 'tax_included',
-        })
-
-        # define a product record with the tax
-        self.env['product.product'].create({
-            'name': 'Test Product',
-            'list_price': 110,
-            'taxes_id': [(6, 0, [tax.id])],
-            'available_in_pos': True,
-        })
-
-        # set Tax-Excluded Price
-        self.main_pos_config.write({
-            'iface_tax_included': 'subtotal'
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'ShowTaxExcludedTour', login="pos_user")
-
-    def test_chrome_without_cash_move_permission(self):
-        self.env.user.write({'group_ids': [
-            Command.set(
-                [
-                    self.env.ref('base.group_user').id,
-                    self.env.ref('point_of_sale.group_pos_user').id,
-                ]
-            )
-        ]})
-        self.main_pos_config.open_ui()
-        self.start_pos_tour('chrome_without_cash_move_permission', login="accountman")
-
-    def test_09_pos_barcodes_scan_product_packaging(self):
-        pack_of_10 = self.env['uom.uom'].create({
-            'name': 'Pack of 10',
-            'relative_factor': 10,
-            'relative_uom_id': self.env.ref('uom.product_uom_unit').id,
-            'is_pos_groupable': True,
-        })
-        product = self.env['product.product'].create({
-            'name': 'Packaging Product',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-            'barcode': '12345601',
-            'uom_ids': [Command.link(pack_of_10.id)],
-        })
-        self.env['product.uom'].create({
-            'barcode': '12345610',
-            'product_id': product.id,
-            'uom_id': pack_of_10.id,
-        })
-        # product has no barcode, it's variant's packaging has it
-        product_templ_2 = self.env['product.template'].create({
-            'name': 'Packaging Product2',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-        })
-        color_attribute = self.env['product.attribute'].create({
-            'name': 'Color always',
-            'create_variant': 'always',
-            'value_ids': [(0, 0, {
-                'name': 'Red',
-                'sequence': 1,
-            }), (0, 0, {
-                'name': 'Blue',
-                'sequence': 2,
-            })],
-        })
-
-        self.env['product.template.attribute.line'].create([
-            {
-                'product_tmpl_id': product_templ_2.id,
-                'attribute_id': color_attribute.id,
-                'value_ids': [(6, 0, color_attribute.value_ids.ids)]
-            },
-            {
-                'product_tmpl_id': product_templ_2.id,
-                'attribute_id': self.chair_addons_attribute.id,
-                'value_ids': [(6, 0, [self.chair_addon_cushion.id, self.chair_addon_cupholder.id])]
-            }
-        ])
-        self.env['product.uom'].create([
-            {
-                'barcode': '12345618',
-                'product_id': product_templ_2.product_variant_ids[0].id,
-                'uom_id': pack_of_10.id,
-            },
-            {
-                'barcode': '12345619',
-                'product_id': product_templ_2.product_variant_ids[1].id,
-                'uom_id': pack_of_10.id,
-            }
-        ])
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'BarcodeScanningProductPackagingTour', login="pos_user")
-
-    def test_GS1_pos_barcodes_scan(self):
-        barcodes_gs1_nomenclature = self.env.ref("barcodes_gs1_nomenclature.default_gs1_nomenclature")
-        default_nomenclature_id = self.env.ref("barcodes.default_barcode_nomenclature")
-        self.main_pos_config.company_id.write({
-            'nomenclature_id': barcodes_gs1_nomenclature.id
-        })
-        self.main_pos_config.write({
-            'fallback_nomenclature_id': default_nomenclature_id
-        })
-        self.env['product.product'].create({
-            'name': 'Product 1',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-            'barcode': '08431673020125',
-        })
-
-        self.env['product.product'].create({
-            'name': 'Product 2',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-            'barcode': '08431673020126',
-        })
-
-        # 3760171283370 can be parsed with GS1 rules but it's not GS1
-        self.env['product.product'].create({
-            'name': 'Product 3',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-            'barcode': '3760171283370',
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'GS1BarcodeScanningTour', login="pos_user")
-
-    def test_gs1_barcode_scan_missing_product_variant(self):
-        """
-        Scanning a GS1 barcode for a product that is not loaded must add the specific
-        matching variant, not the first variant of the template.
-        """
-        barcodes_gs1_nomenclature = self.env.ref("barcodes_gs1_nomenclature.default_gs1_nomenclature")
-        default_nomenclature_id = self.env.ref("barcodes.default_barcode_nomenclature")
-        self.main_pos_config.company_id.write({
-            'nomenclature_id': barcodes_gs1_nomenclature.id,
-        })
-        self.main_pos_config.write({
-            'fallback_nomenclature_id': default_nomenclature_id,
-        })
-
-        size_attribute = self.env['product.attribute'].create({
-            'name': 'Size',
-            'create_variant': 'always',
-            'value_ids': [
-                Command.create({'name': 'L', 'sequence': 1}),
-                Command.create({'name': 'S', 'sequence': 2}),
-            ],
-        })
-        product_tmpl = self.env['product.template'].create({
-            'name': 'GS1 Missing Variant Product',
-            'available_in_pos': False,
-            'list_price': 10,
-            'taxes_id': False,
-            'attribute_line_ids': [Command.create({
-                'attribute_id': size_attribute.id,
-                'value_ids': [Command.set(size_attribute.value_ids.ids)],
-            })],
-        })
-
-        variant_s = product_tmpl.product_variant_ids.filtered(
-            lambda v: any(val.name == 'S' for val in v.product_template_attribute_value_ids.mapped('product_attribute_value_id'))
-        )
-        variant_s.write({'barcode': '5400000002649'})
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'test_gs1_barcode_scan_missing_product_variant', login="pos_user")
-
-    def test_refund_order_with_fp_tax_included(self):
-        # create a fiscal position
-        self.fiscal_position = self.env['account.fiscal.position'].create({
-            'name': 'No Tax',
-        })
-        #create a tax of 15% tax included
-        self.tax1 = self.env['account.tax'].create({
-            'name': 'Tax 1',
-            'amount': 15,
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-            'price_include_override': 'tax_included',
-        })
-        #create a tax of 0%
-        self.tax2 = self.env['account.tax'].create({
-            'name': 'Tax 2',
-            'amount': 0,
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-            'price_include_override': 'tax_included',
-            'fiscal_position_ids': self.fiscal_position,
-            'original_tax_ids': self.tax1,
-        })
-
-        self.product_test = self.env['product.product'].create({
-            'name': 'Product Test',
-            'is_storable': True,
-            'available_in_pos': True,
-            'list_price': 100,
-            'taxes_id': [(6, 0, self.tax1.ids)],
-        })
-
-        #add the fiscal position to the PoS
-        self.main_pos_config.write({
-            'fiscal_position_ids': [(4, self.fiscal_position.id)],
-            'tax_regime_selection': True,
-            })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'FiscalPositionNoTaxRefund', login="pos_user")
-        order = self.env['pos.order'].search([])
-        self.assertTrue(order[0].name == order[1].name + " REFUND")
-
-    def test_customer_display_popup(self):
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'customer_display_shows_qr_popup', login="pos_user")
-
-    def test_printed_receipt_tour(self):
-        self.main_pos_config.write({
-            'basic_receipt': True,
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour("point_of_sale.test_printed_receipt_tour")
-
-    def test_receipt_logo_config(self):
-        self.img_company_b64 = 'iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mP8z8BQDwAEhQGAhKmMIQAAAABJRU5ErkJggg=='
-        self.img_custom_b64 = 'iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mnk5+dQAwAEsgKVw6M+4wAAAABJRU5ErkJggg=='
-        company = self.main_pos_config.company_id
-        company.write({
-            'phone': '123456789',
-            'logo': self.img_company_b64,
-        })
-
-        self.start_pos_tour('point_of_sale.test_receipt_company_logo_tour')
-
-        self.main_pos_config.current_session_id.close_session_from_ui()
-        company.write({'logo': False})
-
-        self.start_pos_tour('point_of_sale.test_receipt_no_logo_tour')
-
-        self.main_pos_config.current_session_id.close_session_from_ui()
-        company.write({'logo': self.img_company_b64})
-        self.main_pos_config.write({
-            'custom_logo': self.img_custom_b64,
-            'use_custom_receipt_info': True,
-            'custom_phone': '555-999',
-        })
-
-        self.start_pos_tour('point_of_sale.test_receipt_custom_logo_tour')
-
-    def test_limited_product_pricelist_loading(self):
-        self.env['ir.config_parameter'].sudo().set_int('point_of_sale.limited_product_count', 1)
-
-        limited_category = self.env['pos.category'].create({
-            'name': 'Limited Category',
-        })
-        product_1 = self.env['product.product'].create({
-            'name': 'Test Product 1',
-            'list_price': 100,
-            'barcode': '0100100',
-            'taxes_id': False,
-            'pos_categ_ids': [(4, limited_category.id)],
-            'available_in_pos': True,
-        })
-
-        self.env['product.product'].create({
-            'name': 'Test Product 3',
-            'list_price': 300,
-            'barcode': '0100300',
-            'taxes_id': False,
-            'pos_categ_ids': [(4, limited_category.id)],
-            'available_in_pos': True,
-        })
-
-        pricelist_item = self.env['product.pricelist.item'].create([{
-            'applied_on': '3_global',
-            'fixed_price': 50,
-        }, {
-            'applied_on': '0_product_variant',
-            'product_id': product_1.id,
-            'fixed_price': 80,
-            'min_quantity': 1,
-        }, {
-            'applied_on': '0_product_variant',
-            'product_id': product_1.id,
-            'fixed_price': 70,
-            'min_quantity': 2,
-        }])
-        self.main_pos_config.write({
-            'iface_available_categ_ids': [],
-            'limit_categories': True,
-        })
-        self.main_pos_config.pricelist_id.write({'item_ids': [(6, 0, pricelist_item.ids)]})
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'limitedProductPricelistLoading', login="pos_user")
-
-    def test_restricted_categories_combo_product(self):
-        """
-        Ensure combo choices product are always loaded if parent is in allowed categories, even when restricted categories are configured:
-        - These combo choices should be visible when configuring the parent combo product but not be visible as product that we can directly sell inside POS
-        - These combo choices should appear on the preparation ticket changes
-        """
-        pos_restricted_categ = self.env["pos.category"].create({
-            "name": "Restricted product",
-        })
-        pos_other_categ = self.env["pos.category"].create({
-            "name": "Other products",
-        })
-        self.env['pos.printer'].create({
-            'name': 'Printer',
-            'printer_type': 'epson_epos',
-            'printer_ip': '0.0.0.0',
-            'use_type': 'preparation',
-            'product_categories_ids': [Command.set(self.env['pos.category'].search([]).ids)],
-        })
-
-        self.main_pos_config.write({
-            'use_order_printer': True,
-            'preparation_printer_ids': [Command.set(self.env['pos.printer'].search([('use_type', '=', 'preparation')]).ids)],
-        })
-        self.main_pos_config.write({
-            "limit_categories": True,
-            "iface_available_categ_ids": [(6, 0, [pos_restricted_categ.id])],
-        })
-        setup_product_combo_items(self)
-        self.office_combo.pos_categ_ids = [(6, 0, [pos_restricted_categ.id])]
-        self.office_combo.combo_ids = [(6, 0, [self.desks_combo.id])]
-        self.desks_combo.combo_item_ids[0].product_id.pos_categ_ids = [(6, 0, [pos_restricted_categ.id])]
-        self.desks_combo.combo_item_ids[1].product_id.pos_categ_ids = [(6, 0, [pos_other_categ.id])]
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_restricted_categories_combo_product', login="pos_user")
-
-    def test_printer_restricts_to_allowed_categories_for_combo(self):
-        setup_product_combo_items(self)
-        self.printer.write({
-            'use_type': 'preparation',
-            'product_categories_ids': [Command.set(self.env['pos.category'].search([('name', '=', 'Category 2')]).ids)],
-        })
-        self.office_combo.write({
-            'pos_categ_ids': [Command.set(self.env['pos.category'].search([('name', '=', 'Category 1')]).ids)],
-        })
-        self.main_pos_config.write({
-            'use_order_printer': True,
-            'preparation_printer_ids': [Command.set(self.printer.ids)],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_printer_restricts_to_allowed_categories_for_combo', login="pos_user")
-
-    def test_printer_not_linked_to_any_combo_category(self):
-        setup_product_combo_items(self)
-        new_category = self.env['pos.category'].create({
-            'name': 'New Category',
-        })
-        self.wall_shelf.write({
-            'pos_categ_ids': [Command.set(new_category.ids)],
-        })
-        self.printer.write({
-            'use_type': 'preparation',
-            'product_categories_ids': [Command.set(new_category.ids)],
-        })
-        self.main_pos_config.write({
-            'use_order_printer': True,
-            'preparation_printer_ids': [Command.set(self.printer.ids)],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_printer_not_linked_to_any_combo_category', login="pos_user")
-
-    def test_multi_product_options(self):
-        product_a = self.env['product.product'].create({
-            'name': 'Product A',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-        })
-
-        chair_multi_attribute = self.env['product.attribute'].create({
-            'name': 'Multi',
-            'display_type': 'multi',
-            'create_variant': 'no_variant',
-        })
-        chair_multi_value_1 = self.env['product.attribute.value'].create({
-            'name': 'Value 1',
-            'attribute_id': chair_multi_attribute.id,
-        })
-        chair_multi_value_2 = self.env['product.attribute.value'].create({
-            'name': 'Value 2',
-            'attribute_id': chair_multi_attribute.id,
-        })
-        self.chair_multi_line = self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product_a.product_tmpl_id.id,
-            'attribute_id': chair_multi_attribute.id,
-            'value_ids': [(6, 0, [chair_multi_value_1.id, chair_multi_value_2.id])]
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'MultiProductOptionsTour', login="pos_user")
-
-    def test_translate_product_name(self):
-        self.env['res.lang']._activate_lang('fr_FR')
-        self.pos_user.write({'lang': 'fr_FR'})
-
-        product = self.env['product.product'].create({
-            'name': 'Test Product',
-            'list_price': 100,
-            'taxes_id': False,
-            'available_in_pos': True,
-        })
-        product.update_field_translations('name', {'fr_FR': 'Testez le produit'})
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'TranslateProductNameTour', login="pos_user")
-
-    def test_properly_display_price(self):
-        """Make sure that when the decimal separator is a comma, the shown orderline price is correct.
-        """
-        lang = self.env['res.lang'].search([('code', '=', self.pos_user.lang)])
-        lang.write({'thousands_sep': '.', 'decimal_point': ','})
-
-        self.env['product.product'].create({
-            'name': 'Test Product',
-            'list_price': 1_453.53,
-            'taxes_id': False,
-            'available_in_pos': True,
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, "DecimalCommaOrderlinePrice", login="pos_user")
-
-    def test_res_partner_scan_barcode(self):
-        # default Customer Barcodes pattern is '042'
-        self.env['res.partner'].create({
-            'name': 'John Doe',
-            'barcode': '0421234567890',
-        })
-        self.env['res.partner'].create({
-            'name': 'Dusty-Bun',
-            'barcode': '0241234567890',
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'BarcodeScanPartnerTour', login="pos_user")
-
-    def test_allow_order_modification_after_validation_error(self):
-        """
-        User error as a result of validation should block the order.
-        Taking action by order modification should be allowed.
-        """
-
-        self.env['product.product'].create({
-            'name': 'Test Product',
-            'list_price': 10.00,
-            'taxes_id': False,
-            'available_in_pos': True,
-        })
-
-        def sync_from_ui_patch(*_args, **_kwargs):
-            raise UserError('Test Error')
-
-        with patch.object(self.env.registry.models['pos.order'], "sync_from_ui", sync_from_ui_patch):
-            # If there is problem in the tour, remove the log catcher to debug.
-            with self.assertLogs(level="WARNING") as log_catcher:
-                self.main_pos_config.with_user(self.pos_user).open_ui()
-                self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'OrderModificationAfterValidationError', login="pos_user")
-
-            warning_outputs = [o for o in log_catcher.output if 'WARNING' in o]
-            self.assertEqual(len(warning_outputs), 1, "Exactly one warning should be logged")
-
-    def test_customer_display(self):
-        self.start_tour(f"/pos_customer_display/{self.main_pos_config.id}/{self.main_pos_config.access_token}?access_token={self.main_pos_config.access_token}", 'CustomerDisplayTour', login="pos_user")
-
-    def test_customer_display_scroll(self):
-        self.start_tour(f"/pos_customer_display/{self.main_pos_config.id}/{self.main_pos_config.access_token}?access_token={self.main_pos_config.access_token}", 'CustomerDisplayTourScroll', login="pos_user")
-
-    def test_customer_display_with_qr(self):
-        self.start_tour(f"/pos_customer_display/{self.main_pos_config.id}/{self.main_pos_config.access_token}?access_token={self.main_pos_config.access_token}", 'CustomerDisplayTourWithQr', login="pos_user")
-
-    def test_combo_refund_different_qty(self):
-        setup_product_combo_items(self)
-        self.desks_combo.write({
-            'qty_free': 2,
-            'qty_max': 2,
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_combo_refund_different_qty')
-
-    def test_order_refund_flow(self):
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_order_refund_flow')
-        self.assertEqual(self.env['mail.mail'].search_count([('email_to', '=', 'test@narendradamodardasmodi.com')]), 1)
-
-    def test_refund_few_quantities(self):
-        """ Test to check that refund works with quantities of less than 0.5 """
-        self.env['product.product'].create({
-            'name': 'Sugar',
-            'list_price': 3,
-            'taxes_id': False,
-            'available_in_pos': True,
-            'uom_id': self.env.ref('uom.product_uom_kgm').id,
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'RefundFewQuantities', login="pos_user")
-
-    def test_refund_multiple_products_amounts_compliance(self):
-        test_product = self.env['product.product'].create({
-            'name': 'Test Product',
-            'list_price': 10.00,
-            'taxes_id': False,
-            'available_in_pos': True,
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        current_session = self.main_pos_config.current_session_id
-
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'refund_multiple_products_amounts_compliance', login="pos_user")
-
-        refund_order = current_session.order_ids.filtered(lambda order: order.is_refund)
-        self.assertEqual(refund_order.lines[0].price_subtotal, 2 * test_product.list_price)
-        total_cash_payment = sum(current_session.mapped('order_ids.payment_ids').filtered(
-            lambda payment: payment.payment_method_id.type == 'cash').mapped('amount')
-        )
-        current_session.post_closing_cash_details(total_cash_payment)
-        current_session.close_session_from_ui()
-        self.assertEqual(current_session.state, 'closed')
-        report_refund_order, report_order = self.env['report.pos.order'].sudo().search([('order_id', 'in', current_session.order_ids.ids)])
-        self.assertEqual(report_order.margin, 20.0)
-        self.assertEqual(report_refund_order.margin, -20.0)
-        self.assertEqual(report_order.price_total, 20.0)
-        self.assertEqual(report_refund_order.price_total, -20.0)
-
-    def test_product_combo_price(self):
-        """ Check that the combo has the expected price """
-        self.desk_organizer.product_variant_id.write({"lst_price": 7})
-        self.desk_pad.product_variant_id.write({"lst_price": 2.5})
-        self.whiteboard_pen.product_variant_id.write({"lst_price": 1.5})
-
-        combos = self.env["product.combo"].create([
-            {
-                "name": product.name,
-                "combo_item_ids": [
-                    Command.create({
-                        "product_id": product.id, "extra_price": 0
-                    })
-                ]
-            }
-            for product in (self.desk_organizer.product_variant_id, self.desk_pad.product_variant_id, self.whiteboard_pen.product_variant_id)
-        ])
-
-        self.env["product.product"].create(
-            {
-                "available_in_pos": True,
-                "list_price": 7,
-                "standard_price": 10,
-                "name": "Desk Combo",
-                "type": "combo",
-                "taxes_id": False,
-                "combo_ids": [
-                    (6, 0, [combo.id for combo in combos])
-                ],
-            }
-        )
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(f"/pos/ui/{self.main_pos_config.id}", 'ProductComboPriceCheckTour', login="pos_user")
-        order = self.env['pos.order'].search([], limit=1)
-        self.assertEqual(order.lines.filtered(lambda l: l.product_id.type == 'combo').margin, 0)
-        self.assertEqual(order.lines.filtered(lambda l: l.product_id.type == 'combo').margin_percent, 0)
-
-    def test_customer_display_as_public(self):
-        self.main_pos_config.customer_display_bg_img = 'iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAIAAACQd1PeAAAADElEQVR4nGNgYGAAAAAEAAH2FzhVAAAAAElFTkSuQmCC'
-        response = self.url_open(f"/web/image/pos.config/{self.main_pos_config.id}/customer_display_bg_img")
-        self.assertEqual(response.status_code, 200)
-        self.assertTrue('Shop.png' in response.headers['Content-Disposition'])
-
-    def test_customer_all_fields_displayed(self):
-        """
-        Verify that all the field of a partner can be displayed in the partner list.
-        Also verify that all these fields can be searched.
-        """
-        self.env["res.partner"].create({
-            "name": "John Doe",
-            "street": "1 street of astreet",
-            "city": "Acity",
-            "state_id": self.env.ref("base.state_us_30").id,  # Ohio
-            "country_id": self.env.ref("base.us").id,
-            "zip": "26432685463",
-            "phone": "9898989899",
-            "email": "john@doe.com"
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('PosCustomerAllFieldsDisplayed')
-
-    def test_product_combo_change_fp(self):
-        """
-        Verify than when the fiscal position is changed,
-        the price of the combo doesn't change and taxes are well taken into account
-        """
-        fiscal_position = self.env['account.fiscal.position'].create({
-            'name': 'test fp',
-        })
-
-        tax_1 = self.env['account.tax'].create({
-            'name': 'Tax 10%',
-            'amount': 10,
-            'price_include_override': 'tax_included',
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-        })
-
-        self.env['account.tax'].create({
-            'name': 'Tax 5%',
-            'amount': 5,
-            'price_include_override': 'tax_included',
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-            'fiscal_position_ids': [Command.link(fiscal_position.id)],
-            'original_tax_ids': [Command.link(tax_1.id)],
-        })
-
-        setup_product_combo_items(self)
-        self.office_combo.write({'list_price': 50, 'taxes_id': [(6, 0, [tax_1.id])]})
-        for combo in self.office_combo.combo_ids:  # Set the tax to all the products of the combo
-            for item in combo.combo_item_ids:
-                item.product_id.taxes_id = [(6, 0, [tax_1.id])]
-
-        self.main_pos_config.write({
-            'tax_regime_selection': True,
-            'fiscal_position_ids': [(6, 0, [fiscal_position.id])],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(f"/pos/ui/{self.main_pos_config.id}", 'ProductComboChangeFP', login="pos_user")
-
-    def test_product_combo_change_pricelist(self):
-        """
-        Verify than when we change the pricelist, the combo price is updated
-        """
-        setup_product_combo_items(self)
-
-        sale_10_pl = self.env['product.pricelist'].create({
-            'name': 'sale 10%',
-        })
-        self.env['product.pricelist.item'].create({
-            'pricelist_id': sale_10_pl.id,
-            'compute_price': 'percentage',
-            'applied_on': '3_global',
-            'percent_price': 10,
-        })
-
-        self.main_pos_config.write({
-            'available_pricelist_ids': [(4, sale_10_pl.id)],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(f"/pos/ui?config_id={self.main_pos_config.id}", 'ProductComboChangePricelist', login="pos_user")
-
-    def test_product_combo_discount(self):
-        """
-        Verify that the combo product applies the correct discount and updates prices accordingly
-        """
-        setup_product_combo_items(self)
-        self.office_combo.write({'list_price': 100})
-        self.office_combo.combo_ids.combo_item_ids.product_id.taxes_id = False
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(
-            f"/pos/ui?config_id={self.main_pos_config.id}",
-            'ProductComboDiscountTour',
-            login="pos_user",
-        )
-
-    def test_combo_item_image_display(self):
-        """ when `show_product_images` is enabled, verify combo item product images should appear in the POS UI. When disabled, the UI should not display
-            the product image for combo items.
-        """
-
-        setup_product_combo_items(self)
-        image = _create_image(color="orange")
-
-        for combo in self.office_combo.combo_ids:
-            for item in combo.combo_item_ids:
-                item.product_id.image_1920 = image
-
-        # Case 1: Images ON → images must be visible in combo items
-        self.main_pos_config.show_product_images = True
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(f"/pos/ui?config_id={self.main_pos_config.id}", 'test_combo_item_image_display', login="pos_user")
-
-        # Case 2: Images OFF → product images should not be shown
-        self.main_pos_config.show_product_images = False
-        self.start_tour(f"/pos/ui?config_id={self.main_pos_config.id}", 'test_combo_item_image_not_display', login="pos_user")
-
-    def test_product_categories_order(self):
-        """ Verify that the order of categories doesnt change in the frontend """
-        self.env['pos.category'].search([]).write({'sequence': 100})
-        aaa_catg = self.env['pos.category'].create({
-            'name': 'AAA',
-            'parent_id': False,
-            'sequence': 1,
-        })
-        aac_catg = self.env['pos.category'].create({
-            'name': 'AAC',
-            'parent_id': False,
-            'sequence': 3,
-        })
-        parentA = self.env['pos.category'].create({
-            'name': 'AAB',
-            'parent_id': False,
-            'sequence': 2,
-        })
-        parentB = self.env['pos.category'].create({
-            'name': 'AAX',
-            'parent_id': parentA.id,
-        })
-        aay_catg = self.env['pos.category'].create({
-            'name': 'AAY',
-            'parent_id': parentB.id,
-        })
-        # Add a product that belongs to both parent and child categories.
-        # It's presence is checked during the tour to make sure app doesn't crash.
-        self.env['product.product'].create({
-            'name': 'Product in AAB and AAX',
-            'pos_categ_ids': [(6, 0, [parentA.id, parentB.id])],
-            'available_in_pos': True,
-        })
-        self.env['product.product'].create([
-            {
-                'name': 'Product in AAA Catg',
-                'pos_categ_ids': [(6, 0, [aaa_catg.id])],
-                'available_in_pos': True,
-            },
-            {
-                'name': 'Product in AAC Catg',
-                'pos_categ_ids': [(6, 0, [aac_catg.id])],
-                'available_in_pos': True,
-            },
-            {
-                'name': 'Product in AAY Catg',
-                'pos_categ_ids': [(6, 0, [aay_catg.id])],
-                'available_in_pos': True,
-            },
-        ])
-        self.main_pos_config.with_user(self.pos_admin).open_ui()
-        self.start_tour(f"/pos/ui/{self.main_pos_config.id}", 'PosCategoriesOrder', login="pos_admin")
-
-    def test_product_with_dynamic_attributes(self):
-        dynamic_attribute = self.env['product.attribute'].create({
-            'name': 'Dynamic Attribute',
-            'create_variant': 'dynamic',
-        })
-        value_1 = self.env['product.attribute.value'].create({
-            'name': 'Test 1',
-            'attribute_id': dynamic_attribute.id,
-        })
-        value_2 = self.env['product.attribute.value'].create({
-            'name': 'Test 2',
-            'default_extra_price': 10,
-            'attribute_id': dynamic_attribute.id,
-        })
-        product_template = self.env['product.template'].create({
-            'name': 'Dynamic Product',
-            'uom_id': self.env.ref('uom.product_uom_unit').id,
-            'is_storable': True,
-            'available_in_pos': True,
-        })
-        self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product_template.id,
-            'attribute_id': dynamic_attribute.id,
-            'value_ids': [Command.set([value_1.id, value_2.id])],
-        })
-        self.main_pos_config.with_user(self.pos_admin).open_ui()
-        self.start_tour(f"/pos/ui/{self.main_pos_config.id}", 'PosProductWithDynamicAttributes', login="pos_admin")
-
-    def test_product_with_single_value_dynamic_attribute(self):
-        """A dynamic attribute with a single value must not open the configurator but still
-        creates the product variant on the server when added to the order."""
-        dynamic_attribute = self.env['product.attribute'].create({
-            'name': 'Single Dynamic Attribute',
-            'create_variant': 'dynamic',
-        })
-        value = self.env['product.attribute.value'].create({
-            'name': 'Only Value',
-            'attribute_id': dynamic_attribute.id,
-        })
-        product_template = self.env['product.template'].create({
-            'name': 'Single Dynamic Product',
-            'list_price': 5.0,
-            'taxes_id': False,
-            'available_in_pos': True,
-        })
-        self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product_template.id,
-            'attribute_id': dynamic_attribute.id,
-            'value_ids': [Command.set([value.id])],
-        })
-
-        no_variant_attribute = self.env['product.attribute'].create({
-            'name': 'No Variant Attribute',
-            'create_variant': 'no_variant',
-        })
-        no_variant_value = self.env['product.attribute.value'].create({
-            'name': 'No Variant Value',
-            'attribute_id': no_variant_attribute.id,
-        })
-        mixed_template = self.env['product.template'].create({
-            'name': 'Mixed Attribute Product',
-            'list_price': 7.0,
-            'taxes_id': False,
-            'available_in_pos': True,
-        })
-        self.env['product.template.attribute.line'].create([
-            {
-                'product_tmpl_id': mixed_template.id,
-                'attribute_id': dynamic_attribute.id,
-                'value_ids': [Command.set([value.id])],
-            },
-            {
-                'product_tmpl_id': mixed_template.id,
-                'attribute_id': no_variant_attribute.id,
-                'value_ids': [Command.set([no_variant_value.id])],
-            },
-        ])
-        self.main_pos_config.with_user(self.pos_admin).open_ui()
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_product_with_single_value_dynamic_attribute', login="pos_user")
-
-    def test_autofill_cash_count(self):
-        """Make sure that when the decimal separator is a comma, the shown orderline price is correct.
-        """
-        lang = self.env['res.lang'].search([('code', '=', self.pos_user.lang)])
-        lang.write({'thousands_sep': '.', 'decimal_point': ','})
-        self.env["product.product"].create(
-            {
-                "available_in_pos": True,
-                "list_price": 123456,
-                "name": "Test Expensive",
-                "taxes_id": False
-            }
-        )
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, "AutofillCashCount", login="pos_user")
-
-    def test_product_search_2(self):
-        self.env['product.product'].create({
-            'name': 'Test chair 1',
-            'available_in_pos': True,
-        })
-        self.env['product.product'].create({
-            'name': 'Test CHAIR 2',
-            'available_in_pos': True,
-        })
-        self.env['product.product'].create({
-            'name': 'Test sofa',
-            'available_in_pos': True,
-            "default_code": "CHAIR_01",
-        })
-        self.env['product.product'].create({
-            'name': 'clémentine',
-            'available_in_pos': True,
-        })
-        self.main_pos_config.open_ui()
-        self.start_tour(f"/pos/ui/{self.main_pos_config.id}", 'SearchProducts', login="pos_user")
-
-    def test_product_search(self):
-        """Verify that the product search works correctly"""
-        product_with_variant = self.env['product.template'].create({
-            'name': 'Product with Variant',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-            'barcode': '1234567',
-        })
-
-        color_attribute = self.env['product.attribute'].create({
-            'name': 'Color always',
-            'create_variant': 'always',
-            'value_ids': [(0, 0, {
-                'name': 'Red',
-                'sequence': 1,
-            }), (0, 0, {
-                'name': 'Blue',
-                'sequence': 2,
-            })],
-        })
-
-        self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product_with_variant.id,
-            'attribute_id': color_attribute.id,
-            'value_ids': [(6, 0, color_attribute.value_ids.ids)]
-        })
-        product_with_variant.product_variant_ids[0].write({
-            "barcode": "variant_barcode_1",
-            "default_code": "VARIANT_1"
-        })
-        product_with_variant.product_variant_ids[1].write({
-            "barcode": "variant_barcode_2",
-            "default_code": "VARIANT_2"
-        })
-
-        self.env['product.product'].create([
-            {
-                'name': 'Test Product 1',
-                'list_price': 100,
-                'taxes_id': False,
-                'available_in_pos': True,
-                'barcode': '1234567890123',
-                'default_code': 'TESTPROD1',
-            },
-            {
-                'name': 'Test Product 2',
-                'list_price': 100,
-                'taxes_id': False,
-                'available_in_pos': True,
-                'barcode': '1234567890124',
-                'default_code': 'TESTPROD2',
-            },
-            {
-                'name': 'Apple',
-                'list_price': 100,
-                'taxes_id': False,
-                'available_in_pos': True,
-            },
-            {
-                'name': 'galaxy',
-                'list_price': 100,
-                'taxes_id': False,
-                'available_in_pos': True,
-            },
-            {
-                'name': '1234567890123',
-                'list_price': 100,
-                'taxes_id': False,
-                'available_in_pos': True,
-            },
-        ])
-
-        att_color = self.env['product.attribute'].create({'name': 'Color', 'sequence': 1})
-
-        att_color_values = self.env['product.attribute.value'].create([
-            {'name': 'galaxy variant', 'attribute_id': att_color.id, 'sequence': 1},
-            {'name': 'blue', 'attribute_id': att_color.id, 'sequence': 2},
-            ])
-
-        self.env['product.template'].create({
-            'name': 'Test Product variant',
-            'attribute_line_ids': [
-                Command.create({
-                    'attribute_id': att_color.id,
-                    'value_ids': [Command.set(att_color_values.mapped('id'))],
-                }),
-            ],
-            'available_in_pos': True,
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'ProductSearchTour', login="pos_user")
-
     def test_customer_popup(self):
         """Verify that the customer popup search & inifnite scroll work properly"""
         self.env["res.partner"].create([{"name": "Z partner to search"}, {"name": "Z partner to scroll"}])
         self.main_pos_config.with_user(self.pos_user).open_ui()
         self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'CustomerPopupTour', login="pos_user")
-
-    def test_pricelist_multi_items_different_qty_thresholds(self):
-        """ Having multiple pricelist items for the same product tmpl with ascending `min_quantity`
-        values, prefer the "latest available"- that is, the one with greater `min_quantity`.
-        """
-        product = self.env['product.product'].create({
-            'name': 'tpmcapi product',
-            'list_price': 1.0,
-            'available_in_pos': True,
-            'taxes_id': False,
-        })
-        self.main_pos_config.pricelist_id.write({
-            'item_ids': [Command.create({
-                'display_applied_on': '1_product',
-                'product_tmpl_id': product.product_tmpl_id.id,
-                'compute_price': 'fixed',
-                'fixed_price': 10.0,
-                'min_quantity': 3,
-            }), Command.create({
-                'display_applied_on': '1_product',
-                'product_tmpl_id': product.product_tmpl_id.id,
-                'compute_price': 'fixed',
-                'fixed_price': 20.0,
-                'min_quantity': 2,
-            })],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(
-            f'/pos/ui/{self.main_pos_config.id}',
-            'test_pricelist_multi_items_different_qty_thresholds',
-            login='pos_user'
-        )
-
-    def test_tracking_number_closing_session(self):
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(f"/pos/ui/{self.main_pos_config.id}", 'test_tracking_number_closing_session', login="pos_user")
-
-        # Change should be given in cash
-        cash_payment_method = self.main_pos_config.payment_method_ids.filtered(lambda p: p.is_cash_count)
-        last_order = self.main_pos_config.current_session_id.order_ids[-1]
-        self.assertRecordValues(last_order.payment_ids.sorted(), [
-            {'amount': -18.02, 'payment_method_id': cash_payment_method.id, 'is_change': True},
-            {'amount': 20.0, 'payment_method_id': self.bank_payment_method.id, 'is_change': False},
-        ])
-
-        # References should not have gaps
-        references = self.env['pos.order'].search([], order="pos_reference").mapped("pos_reference")
-        for i in range(len(references) - 1):
-            self.assertEqual(int(references[i + 1].split('-')[-1]), int(references[i].split('-')[-1]) + 1, "There is a gap in the pos references")
-
-    def test_reload_page_before_payment_with_customer_account(self):
-        self.customer_account_payment_method = self.env['pos.payment.method'].create({
-            'name': 'Customer Account',
-            'split_transactions': True,
-        })
-        self.main_pos_config.write({'payment_method_ids': [(6, 0, self.customer_account_payment_method.ids)]})
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(
-            f'/pos/ui/{self.main_pos_config.id}',
-            'test_reload_page_before_payment_with_customer_account',
-            login='pos_user'
-        )
-
-    def test_product_card_qty_precision(self):
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(f"/pos/ui/{self.main_pos_config.id}", 'ProductCardUoMPrecision', login="pos_user")
-
-    @freeze_time("2025-06-15 11:09")
-    def test_cash_in_out(self):
-        self.main_pos_config.with_user(self.pos_admin).open_ui()
-        self.start_tour(f"/pos/ui/{self.main_pos_config.id}", 'test_cash_in_out', login="pos_admin")
-
-        self.assertEqual(len(self.main_pos_config.current_session_id.statement_line_ids), 1, "There should be one cash in/out statement line")
-        self.assertEqual(self.main_pos_config.current_session_id.statement_line_ids[0].amount, -5, "The cash in/out amount should be -5")
-
-    def test_edit_paid_order(self):
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(f"/pos/ui/{self.main_pos_config.id}", 'test_edit_paid_order', login="pos_user")
-        edited_orders = self.env['pos.order'].search([], limit=1)
-        # check invoice created
-        self.assertTrue(edited_orders[0].account_move)
-        self.assertEqual(edited_orders[0].partner_id.name, 'Partner Test 1')
-
-    def test_reuse_empty_floating_order(self):
-        """ Verify that after a payment, POS should reuse an existing empty floating order if available, instead of always creating new ones """
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(f"/pos/ui?config_id={self.main_pos_config.id}", 'test_reuse_empty_floating_order', login="pos_user")
-
-    def test_order_and_invoice_amounts(self):
-        payment_term = self.env['account.payment.term'].create({
-            'name': "early_payment_term",
-            'discount_percentage': 10,
-            'discount_days': 10,
-            'early_discount': True,
-            'early_pay_discount_computation': 'mixed',
-            'line_ids': [Command.create({
-                'value': 'percent',
-                'nb_days': 0,
-                'value_amount': 100,
-            })]
-        })
-        self.partner_test_1.property_payment_term_id = payment_term.id
-
-        tax = self.env['account.tax'].create({
-            'name': 'Tax 10%',
-            'amount': 10,
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-        })
-        self.env['product.product'].create({
-            'name': 'Product Test',
-            'available_in_pos': True,
-            'list_price': 1000,
-            'taxes_id': [(6, 0, [tax.id])],
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'PaymentScreenInvoiceOrder', login="pos_user")
-
-        order = self.env['pos.order'].search([('partner_id', '=', self.partner_test_1.id)], limit=1)
-        self.assertTrue(order)
-
-        self.assertEqual(order.partner_id, self.partner_test_1)
-
-        invoice = self.env['account.move'].search([('invoice_origin', '=', order.pos_reference)], limit=1)
-        self.assertTrue(invoice)
-        self.assertFalse(invoice.invoice_payment_term_id)
-
-        self.assertAlmostEqual(order.amount_total, invoice.amount_total, places=2, msg="Order and Invoice amounts do not match.")
-
-    def test_pricelist_parent_category_rule(self):
-        parent_category = self.env['product.category'].create({
-            'name': 'Parent Category',
-        })
-        child_category = self.env['product.category'].create({
-            'name': 'Child Category',
-            'parent_id': parent_category.id,
-        })
-        self.env['product.product'].create({
-            'name': 'Product with child category',
-            'list_price': 100,
-            'taxes_id': False,
-            'available_in_pos': True,
-            'categ_id': child_category.id,
-        })
-
-        pricelist = self.env['product.pricelist'].create({
-            'name': 'Test pricelist on category',
-            'item_ids': [(0, 0, {
-                'compute_price': 'fixed',
-                'fixed_price': 50,
-                'applied_on': '2_product_category',
-                'categ_id': parent_category.id,
-            })],
-        })
-
-        self.main_pos_config.write({
-            'pricelist_id': pricelist.id,
-            'available_pricelist_ids': [(6, 0, [pricelist.id])],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour(f"/pos/ui?config_id={self.main_pos_config.id}", 'test_pricelist_parent_category_rule', login="pos_user")
-
-    def test_product_create_update_from_frontend(self):
-        ''' This test verifies product creation and updates product details from the POS frontend. '''
-        self.pos_admin.write({
-            'group_ids': [Command.link(self.env.ref('base.group_system').id)],
-        })
-        self.env['pos.category'].search([('id', '!=', self.pos_cat_chair_test.id)]).write({'sequence': 100})
-        self.pos_cat_chair_test.write({'sequence': 1})
-        self.main_pos_config.with_user(self.pos_admin).open_ui()
-        self.start_tour('/pos/ui/%d' % self.main_pos_config.id, 'test_product_create_update_from_frontend', login='pos_admin')
-
-        # In the frontend, a product was created during the tour with the following details:
-        # - Product name: Test Frontend Product
-        # - Barcode: 710535977349
-        # - List price: 20.0
-
-        #  Ensure that the original product created in the frontend ('Test Frontend Product') has been edited to ('Test Frontend Product Edited').
-        frontend_created_product = self.env['product.product'].search_count([('name', '=', 'Test Frontend Product')])
-        frontend_created_product_edited = self.env['product.product'].search([('name', '=', 'Test Frontend Product Edited')])
-
-        self.assertEqual(frontend_created_product, 0)
-        self.assertEqual(frontend_created_product_edited.name, 'Test Frontend Product Edited')
-        self.assertEqual(frontend_created_product_edited.barcode, '710535977348')
-        self.assertEqual(frontend_created_product_edited.list_price, 50.0)
-
-    def test_one_attribute_value_scan_barcode(self):
-        product = self.env['product.template'].create({
-            'name': 'Product Test',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-            'barcode': '1234567',
-        })
-
-        size_attribute = self.env['product.attribute'].create({
-            'name': 'Size never',
-            'create_variant': 'no_variant',
-            'value_ids': [(0, 0, {
-                'name': 'Large',
-            })],
-        })
-
-        self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product.id,
-            'attribute_id': size_attribute.id,
-            'value_ids': [(6, 0, size_attribute.value_ids.ids)]
-        })
-
-        color_attribute = self.env['product.attribute'].create({
-            'name': 'Color always',
-            'create_variant': 'always',
-            'value_ids': [(0, 0, {
-                'name': 'Red',
-                'sequence': 1,
-            }), (0, 0, {
-                'name': 'Blue',
-                'sequence': 2,
-            })],
-        })
-
-        self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product.id,
-            'attribute_id': color_attribute.id,
-            'value_ids': [(6, 0, color_attribute.value_ids.ids)]
-        })
-
-        product.product_variant_ids[0].barcode = '1234567'
-        product.product_variant_ids[1].barcode = '1234568'
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_one_attribute_value_scan_barcode', login="pos_user")
-
-    def test_fiscal_position_tax_group_labels(self):
-        fiscal_position = self.env['account.fiscal.position'].create({
-            'name': 'Fiscal Position Test',
-        })
-        tax_1 = self.env['account.tax'].create({
-            'name': 'Tax 15%',
-            'amount': 15,
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-            'tax_group_id': self.env['account.tax.group'].create({
-                'name': 'Tax Group 15%',
-                'company_id': self.env.company.id,
-                'pos_receipt_label': 'Tax Group 1',
-            }).id,
-        })
-
-        tax_2 = self.env['account.tax'].create({
-            'name': 'Tax 5%',
-            'amount': 5,
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-            'tax_group_id': self.env['account.tax.group'].create({
-                'name': 'Tax Group 5%',
-                'company_id': self.env.company.id,
-                'pos_receipt_label': 'Tax Group 2',
-            }).id,
-            'fiscal_position_ids': [Command.link(fiscal_position.id)],
-            'original_tax_ids': [Command.link(tax_1.id)],
-        })
-
-        self.product = self.env['product.product'].create({
-            'name': 'Test Product',
-            'taxes_id': [(6, 0, [tax_1.id])],
-            'list_price': 100,
-            'available_in_pos': True,
-        })
-
-        self.main_pos_config.write({
-            'tax_regime_selection': True,
-            'fiscal_position_ids': [(6, 0, [fiscal_position.id])],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_fiscal_position_tax_group_labels')
-        orders = self.main_pos_config.current_session_id.order_ids
-
-        self.assertEqual(orders[0].fiscal_position_id.id, fiscal_position.id)
-        self.assertEqual(orders[0].lines.tax_ids_after_fiscal_position.id, tax_2.id)
-        self.assertEqual(orders[0].amount_total, 105)
-        self.assertFalse(orders[1].fiscal_position_id)
-        self.assertEqual(orders[1].lines.tax_ids_after_fiscal_position.id, tax_1.id)
-        self.assertEqual(orders[1].amount_total, 115)
-
-    def test_draft_orders_not_syncing(self):
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_draft_orders_not_syncing', login="pos_user")
-        n_draft_order = self.env['pos.order'].search_count([('state', '=', 'draft')], limit=1)
-        self.assertEqual(n_draft_order, 0, 'There should be no draft orders created')
 
     def test_product_long_press(self):
         """ Test the long press on product to open the product info """
@@ -2533,1369 +733,6 @@ class TestUi(TestPointOfSaleHttpCommon):
         })
         self.main_pos_config.with_user(self.pos_user).open_ui()
         self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'test_product_long_press', login="pos_user")
-
-    def test_zero_decimal_places_currency(self):
-        zero_decimal_currency = self.env['res.currency'].create({
-            'name': 'ZeroDecimalCurrency',
-            'symbol': 'ZDC',
-            'rounding': 1.0,
-            'decimal_places': 0,
-        })
-
-        self.env.user.company_id.currency_id = zero_decimal_currency
-        self.main_pos_config.available_pricelist_ids.write({'currency_id': zero_decimal_currency.id})
-
-        self.env['product.product'].create({
-            'name': 'Test Product',
-            'list_price': 100,
-            'taxes_id': False,
-            'available_in_pos': True,
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_zero_decimal_places_currency', login="pos_user")
-        order = self.env['pos.order'].search([], limit=1)
-        self.assertEqual(order.payment_ids[0].payment_method_id.name, "Bank")
-
-    def test_barcode_search_attributes_preset(self):
-        product = self.env['product.template'].create({
-            'name': 'Product with Attributes',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-        })
-
-        # Product template to force UI reset (acts as a delay)
-        self.env['product.template'].create({
-            'name': 'Product without Attributes',
-            'available_in_pos': True,
-            'list_price': 20,
-            'taxes_id': False,
-            'barcode': '987654321',
-        })
-
-        attribute_1, attribute_2, attribute_3, attribute_4 = self.env['product.attribute'].create([{
-            'name': 'Attribute 1',
-            'create_variant': 'always',
-            'display_type': 'radio',
-            'value_ids': [(0, 0, {
-                'name': 'Value 1',
-            }), (0, 0, {
-                'name': 'Value 2',
-            })],
-        }, {
-            'name': 'Attribute 2',
-            'create_variant': 'always',
-            'display_type': 'pills',
-            'value_ids': [(0, 0, {
-                'name': 'Value 3',
-            }), (0, 0, {
-                'name': 'Value 4',
-            })],
-        }, {
-            'name': 'Attribute 3',
-            'create_variant': 'always',
-            'display_type': 'select',
-            'value_ids': [(0, 0, {
-                'name': 'Value 5',
-            }), (0, 0, {
-                'name': 'Value 6',
-            })],
-        }, {
-            'name': 'Attribute 4',
-            'create_variant': 'always',
-            'display_type': 'color',
-            'value_ids': [(0, 0, {
-                'name': 'Value 7',
-            }), (0, 0, {
-                'name': 'Value 8',
-            })],
-        }])
-
-        self.env['product.template.attribute.line'].create([{
-            'product_tmpl_id': product.id,
-            'attribute_id': attribute_1.id,
-            'value_ids': [(6, 0, attribute_1.value_ids.ids)],
-            'sequence': 1,
-        }, {
-            'product_tmpl_id': product.id,
-            'attribute_id': attribute_2.id,
-            'value_ids': [(6, 0, attribute_2.value_ids.ids)],
-            'sequence': 2,
-        }, {
-            'product_tmpl_id': product.id,
-            'attribute_id': attribute_3.id,
-            'value_ids': [(6, 0, attribute_3.value_ids.ids)],
-            'sequence': 3,
-        }, {
-            'product_tmpl_id': product.id,
-            'attribute_id': attribute_4.id,
-            'value_ids': [(6, 0, attribute_4.value_ids.ids)],
-            'sequence': 4,
-        }])
-
-        for p in product.product_variant_ids:
-            p.write({
-                'barcode': f'1234{"".join(p.product_template_attribute_value_ids.mapped(lambda ptav: ptav.name[-1]))}',
-            })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_barcode_search_attributes_preset', login="pos_user")
-
-    def test_auto_validate_force_done(self):
-        self.main_pos_config.write({
-            'auto_validate_electronic_payment': True
-        })
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_auto_validate_force_done', login="pos_user")
-
-    def test_pos_ui_round_globally(self):
-        self.main_pos_config.company_id.tax_calculation_rounding_method = 'round_globally'
-        tax_16 = self.env['account.tax'].create({
-            'name': 'Tax 16%',
-            'amount': 16,
-        })
-        self.env['product.product'].create([{
-            'name': 'Test Product 1',
-            'list_price': 7051.73,
-            'taxes_id': [(6, 0, [tax_16.id])],
-            'available_in_pos': True,
-        }, {
-            'name': 'Test Product 2',
-            'list_price': 352.59,
-            'taxes_id': [(6, 0, [tax_16.id])],
-            'available_in_pos': True,
-        }])
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_pos_ui_round_globally', login="pos_user")
-
-        pos_session = self.main_pos_config.current_session_id
-        self.assertEqual(pos_session.order_ids[0].payment_ids[0].amount, 7771.01)
-
-        # Close the session and check the session journal entry.
-        pos_session.action_pos_session_validate()
-
-        lines = pos_session.move_id.line_ids.sorted('balance')
-
-        self.assertEqual(len(lines), 5, "There should be 5 lines in the session journal entry")
-        self.assertAlmostEqual(lines[0].balance, -7051.73)
-        self.assertAlmostEqual(lines[1].balance, -1128.28)
-        self.assertAlmostEqual(lines[2].balance, 56.41)
-        self.assertAlmostEqual(lines[3].balance, 352.59)
-        self.assertAlmostEqual(lines[4].balance, 7771.01)
-
-    def test_ctrl_number_ignored(self):
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_ctrl_number_ignored', login="pos_user")
-
-    def test_click_all_orders_keep_customer(self):
-        """Verify that clicking on 'All Orders' keeps the customer selected."""
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_click_all_orders_keep_customer', login="pos_user")
-
-    def test_quantity_package_of_non_basic_unit(self):
-        test_uom_unit = self.env['uom.uom'].create({
-            "name": "test unit uom",
-            "relative_factor": "1.0",
-        })
-        pack_of_12_unit = self.env['uom.uom'].create({
-            'name': 'Pack of 12 unit',
-            'relative_factor': 12,
-            'relative_uom_id': test_uom_unit.id,
-            'is_pos_groupable': True,
-        })
-        product_cord = self.env['product.product'].create({
-            'name': 'Cord',
-            'is_storable': True,
-            'available_in_pos': True,
-            'uom_id': test_uom_unit.id,
-            'uom_ids': [pack_of_12_unit.id],
-            'lst_price': 10.0,
-        })
-        self.env['product.uom'].create({
-            'barcode': '555555',
-            'product_id': product_cord.id,
-            'uom_id': pack_of_12_unit.id,
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_quantity_package_of_non_basic_unit', login="pos_user")
-
-    def test_attribute_order(self):
-        product = self.env['product.template'].create({
-            'name': 'Product Test',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-        })
-
-        attribute_3 = self.env['product.attribute'].create({
-            'name': 'Attribute 3',
-            'create_variant': 'no_variant',
-            'value_ids': [(0, 0, {
-                'name': 'Value 3',
-            }), (0, 0, {
-                'name': 'Value 4',
-            })],
-        })
-
-        self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product.id,
-            'attribute_id': attribute_3.id,
-            'value_ids': [(6, 0, attribute_3.value_ids.ids)],
-            'sequence': 3,
-        })
-
-        attribute_2 = self.env['product.attribute'].create({
-            'name': 'Attribute 2',
-            'create_variant': 'no_variant',
-            'value_ids': [(0, 0, {
-                'name': 'Value 2',
-            })],
-        })
-
-        self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product.id,
-            'attribute_id': attribute_2.id,
-            'value_ids': [(6, 0, attribute_2.value_ids.ids)],
-            'sequence': 2,
-        })
-
-        attribute_1 = self.env['product.attribute'].create({
-            'name': 'Attribute 1',
-            'create_variant': 'no_variant',
-            'value_ids': [(0, 0, {
-                'name': 'Value 1',
-            })],
-        })
-
-        self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product.id,
-            'attribute_id': attribute_1.id,
-            'value_ids': [(6, 0, attribute_1.value_ids.ids)],
-            'sequence': 1,
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_attribute_order', login="pos_user")
-
-    def test_preset_timing_retail(self):
-        """
-        Test to set order preset hour inside a tour
-        """
-        self.preset_dine_in = self.env['pos.preset'].create({
-            'name': 'Dine in',
-        })
-        self.preset_delivery = self.env['pos.preset'].create({
-            'name': 'Delivery',
-            'identification': 'address',
-        })
-        self.main_pos_config.write({
-            'use_presets': True,
-            'default_preset_id': self.preset_dine_in.id,
-            'available_preset_ids': [(6, 0, [self.preset_delivery.id])],
-        })
-        self.pos_user.street = 'Rue de Ramillies'
-        resource_calendar = self.env['resource.calendar'].create({
-            'name': 'Takeaway',
-            'attendance_ids': [(0, 0, {
-                'dayofweek': str(day),
-                'hour_from': 0,
-                'hour_to': 24,
-            }) for day in range(7)],
-        })
-        self.preset_delivery.write({
-            'use_timing': True,
-            'resource_calendar_id': resource_calendar
-        })
-        self.start_pos_tour('test_preset_timing_retail')
-
-    def test_pricelists_in_pos(self):
-        pos_limited_category = self.env['pos.category'].create({'name': 'Limited Category'})
-        pos_category = self.env['pos.category'].create({'name': 'test_pricelists_in_pos'})
-        product_category = self.env['product.category'].create({'name': 'test_pricelists_in_pos'})
-        orange_category = self.env['product.category'].create({'name': 'Orange Category'})
-
-        def generate_pricelist_items(pricelist, fixed_price, product=None, product_tmpl=None, product_category=None):
-            applied_on = '0_product_variant' if product else '1_product' if product_tmpl else '2_product_category' if product_category else '3_global'
-            return self.env['product.pricelist.item'].create({
-                'pricelist_id': pricelist.id,
-                'product_id': product.id if product else False,
-                'product_tmpl_id': product_tmpl.id if product_tmpl else False,
-                'categ_id': product_category.id if product_category else False,
-                'compute_price': 'fixed',
-                'applied_on': applied_on,
-                'fixed_price': fixed_price,
-            })
-
-        def generate_product_template_with_attributes(name, price, pos_category=None, product_category=None):
-            size_attribute = self.env['product.attribute'].create({
-                'name': 'Size',
-                'sequence': 4,
-                'value_ids': [(0, 0, {
-                    'name': 'BIG',
-                    'sequence': 1,
-                }), (0, 0, {
-                    'name': 'MEDIUM',
-                    'sequence': 2,
-                }), (0, 0, {
-                    'name': 'SMALL',
-                    'sequence': 3,
-                })],
-            })
-
-            product_tmpl = self.env['product.template'].create({
-                'name': name.capitalize(),
-                'available_in_pos': True,
-                'categ_id': product_category.id if product_category else False,
-                'pos_categ_ids': [(4, pos_category.id)] if pos_category else False,
-                'list_price': price,
-                'taxes_id': False,
-                'attribute_line_ids': [(0, 0, {
-                    'attribute_id': size_attribute.id,
-                    'value_ids': [(6, 0, size_attribute.value_ids.ids)]
-                })],
-            })
-
-            for index, variant in enumerate(product_tmpl.product_variant_ids):
-                variant.write({'barcode': f'{name}_{index}'})
-
-            return product_tmpl
-
-        banana = generate_product_template_with_attributes('banana', 10.00, pos_category)
-        apple = generate_product_template_with_attributes('apple', 5.00, False, product_category)
-        pear = generate_product_template_with_attributes('pear', 2.00)
-        lime = generate_product_template_with_attributes('lime', 1.00)
-        orange = generate_product_template_with_attributes('orange', 3.00, False, orange_category)
-        kiwi = generate_product_template_with_attributes('kiwi', 4.00)
-
-        test_pricelist = self.env['product.pricelist'].create({
-            'name': 'Test Pricelist',
-        })
-
-        percentage_pricelist = self.env['product.pricelist'].create({
-            'name': 'Percentage Pricelist',
-        })
-
-        generate_pricelist_items(test_pricelist, 20, False, banana)
-        generate_pricelist_items(test_pricelist, 100, banana.product_variant_ids[0])
-        generate_pricelist_items(test_pricelist, 150, banana.product_variant_ids[1])
-        generate_pricelist_items(test_pricelist, 500, False, False, product_category)
-        generate_pricelist_items(test_pricelist, 1000, False, False, orange_category)
-        generate_pricelist_items(test_pricelist, 100, apple.product_variant_ids[0])
-        generate_pricelist_items(test_pricelist, 20, pear.product_variant_ids[0])
-        generate_pricelist_items(test_pricelist, 40, pear.product_variant_ids[1])
-        generate_pricelist_items(test_pricelist, 60, pear.product_variant_ids[2])
-        generate_pricelist_items(test_pricelist, 100, False, lime)
-        generate_pricelist_items(test_pricelist, 200, lime.product_variant_ids[1])
-        generate_pricelist_items(test_pricelist, 400, lime.product_variant_ids[2])
-        generate_pricelist_items(test_pricelist, 600, orange.product_variant_ids[1])
-        generate_pricelist_items(test_pricelist, 500, orange.product_variant_ids[2])
-        generate_pricelist_items(test_pricelist, 10)
-        generate_pricelist_items(test_pricelist, 20, kiwi.product_variant_ids[0])
-
-        self.env['product.pricelist.item'].create({
-            'pricelist_id': percentage_pricelist.id,
-            'base': 'pricelist',
-            'base_pricelist_id': test_pricelist.id,
-            'compute_price': 'percentage',
-            'percent_price': 50,
-            'applied_on': '3_global',
-        })
-
-        self.main_pos_config.write({
-            "limit_categories": True,
-            "iface_available_categ_ids": [(6, 0, [pos_limited_category.id])],
-            'available_pricelist_ids': [(6, 0, [test_pricelist.id, percentage_pricelist.id])],
-            'pricelist_id': test_pricelist.id,
-        })
-
-        load_product_from_pos_stats = {'count': 0, 'items': {}}
-        product_template = self.env.registry.models['product.template']
-
-        # Test product exclusion
-        cherry = generate_product_template_with_attributes('cherry', 2.00)
-        color_attribute = self.env['product.attribute'].create({
-            'name': 'Color',
-            'sequence': 5,
-            'value_ids': [(0, 0, {
-                'name': 'RED',
-                'sequence': 1,
-            }), (0, 0, {
-                'name': 'GREEN',
-                'sequence': 2,
-            }), (0, 0, {
-                'name': 'BLUE',
-                'sequence': 3,
-            })],
-        })
-        cherry.attribute_line_ids = [(0, 0, {
-            'attribute_id': color_attribute.id,
-            'value_ids': [(6, 0, color_attribute.value_ids.ids)]
-        })]
-        color_attribute = cherry.attribute_line_ids.filtered(lambda l: l.attribute_id.name == 'Color')
-        first_color_value = color_attribute.product_template_value_ids.filtered(lambda v: v.attribute_id.name == 'Color' and v.name == 'RED')
-        first_size_value = cherry.product_variant_ids.product_template_attribute_value_ids.filtered(lambda v: v.attribute_id.name == 'Size' and v.name == 'BIG')
-        first_color_value.excluded_value_ids = [Command.link(value) for value in first_size_value.ids]
-        for index, variant in enumerate(cherry.product_variant_ids):
-            variant.write({'barcode': f'cherry_{index}'})
-
-        @api.model
-        def load_product_from_pos_patch(self, config_id, domain, offset=0, limit=None):
-            load_product_from_pos_stats['count'] += 1
-            result = super(product_template, self).load_product_from_pos(config_id, domain, offset, limit)
-            lowered_name = result['product.template'][0]['display_name'].lower()
-            load_product_from_pos_stats['items'][lowered_name] = len(result['product.pricelist.item'])
-            return result
-
-        with patch.object(product_template, "load_product_from_pos", load_product_from_pos_patch):
-            self.start_pos_tour('test_pricelists_in_pos')
-
-        # Should load 6 different products, since 6 products were created
-        self.assertEqual(load_product_from_pos_stats['count'], 7)
-
-        # Length of loaded pricelist items should correspond to the number of items linked
-        # to the product template or product variant
-        # Global rules are loaded at starting of the PoS
-        self.assertEqual(load_product_from_pos_stats['items']['banana'], 3, "Banana should have 3 pricelist items")
-        self.assertEqual(load_product_from_pos_stats['items']['apple'], 1, "Apple should have 1 pricelist item")
-        self.assertEqual(load_product_from_pos_stats['items']['pear'], 3, "Pear should have 3 pricelist items")
-        self.assertEqual(load_product_from_pos_stats['items']['lime'], 3, "Lime should have 3 pricelist items")
-        self.assertEqual(load_product_from_pos_stats['items']['orange'], 2, "Orange should have 2 pricelist items")
-        self.assertEqual(load_product_from_pos_stats['items']['kiwi'], 1, "Kiwi should have 1 pricelist item")
-
-    def test_available_children_categories(self):
-        parent_categ = self.env['pos.category'].create({
-            'name': 'Parent Category',
-        })
-        children_categs = self.env['pos.category'].create([{
-            'name': 'Child Category 1',
-            'parent_id': parent_categ.id,
-        }, {
-            'name': 'Child Category 2',
-            'parent_id': parent_categ.id,
-        }])
-        self.env['product.product'].create([{
-            'name': 'parent product',
-            'pos_categ_ids': [(6, 0, [parent_categ.id])],
-            'available_in_pos': True,
-        }, {
-            'name': 'child product 1',
-            'pos_categ_ids': [(6, 0, [parent_categ.id, children_categs[0].id])],
-            'available_in_pos': True,
-        }, {
-            'name': 'child product 2',
-            'pos_categ_ids': [(6, 0, [parent_categ.id, children_categs[1].id])],
-            'available_in_pos': True,
-        }])
-        self.main_pos_config.write({
-            'limit_categories': True,
-            'iface_available_categ_ids': [(6, 0, [parent_categ.id, children_categs[1].id])],
-        })
-        self.main_pos_config.open_ui()
-        loaded_data = self.main_pos_config.current_session_id.load_data([])
-        category_id = [category['id'] for category in loaded_data['pos.category']]
-        self.assertNotIn(children_categs[0].id, category_id, "Child category is unavailable and shouldn't appear in the POS")
-        self.assertIn(children_categs[1].id, category_id, "Child category is available and should appear in the POS")
-
-    def test_available_product_uom_ids(self):
-        # Making sure that all of the non-special products that are included in the `load_data` are the ones created in this method.
-        self.env['product.template'].search([]).write({'is_favorite': False})
-
-        self.env['ir.config_parameter'].sudo().set_str('point_of_sale.limited_product_count', '2')
-        uom = self.env['uom.uom'].create({
-            'name': 'Random UOM',
-            'relative_uom_id': self.env.ref('uom.product_uom_unit').id,
-        })
-        product_one, product_two, product_three = self.env['product.product'].create([{
-            'name': "product_one",
-            'available_in_pos': True,
-            'is_favorite': True,
-        },
-        {
-            'name': "product_two",
-            'available_in_pos': True,
-            'is_favorite': True,
-        },
-        {
-            'name': "product_three",
-            'available_in_pos': True,
-        }])
-
-        _, _, product_uom_three = self.env['product.uom'].create([{
-            'barcode': "product_one_barcode",
-            'uom_id': uom.id,
-            'product_id': product_one.id,
-        },
-        {
-            'barcode': "product_two_barcode",
-            'uom_id': uom.id,
-            'product_id': product_two.id,
-        },
-        {
-            'barcode': "product_three_barcode",
-            'uom_id': uom.id,
-            'product_id': product_three.id,
-        },
-        ])
-
-        self.env['product.template'].flush_model()
-        self.main_pos_config.open_ui()
-        loaded_data = self.main_pos_config.current_session_id.load_data([])
-        loaded_product_uoms = [loaded_product_uom['id'] for loaded_product_uom in loaded_data['product.uom']]
-
-        self.assertNotIn(product_uom_three.id, loaded_product_uoms, f"Product UOM {product_uom_three} shouldn't be loaded as its product {product_three} is not included in the results")
-
-    def test_fast_payment_validation_from_product_screen_without_automatic_receipt_printing(self):
-        self.preset_delivery = self.env['pos.preset'].create({
-            'name': 'Delivery',
-            'identification': 'address',
-        })
-        self.main_pos_config.write({
-            'use_fast_payment': True,
-            'use_presets': True,
-            'fast_payment_method_ids': [(6, 0, self.bank_payment_method.ids)],
-            'default_preset_id': self.preset_delivery.id,
-            'available_preset_ids': [(6, 0, [self.preset_delivery.id])],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_fast_payment_validation_from_product_screen_without_automatic_receipt_printing')
-        order1 = self.main_pos_config.current_session_id.order_ids[0]
-        order2 = self.main_pos_config.current_session_id.order_ids[1]
-        self.assertEqual(order1.state, 'paid', "The order should be paid after the fast payment validation")
-        self.assertEqual(len(order1.payment_ids), 1, "There should be one payment line used for the fast payment")
-        self.assertEqual(order1.payment_ids.payment_method_id, self.bank_payment_method, "The payment method used should be the bank payment method")
-        self.assertEqual(order2.state, 'paid', "The order should be paid")
-        self.assertEqual(len(order2.payment_ids), 1, "There should be one payment line")
-        self.assertEqual(order2.payment_ids.payment_method_id, self.bank_payment_method, "The payment method used should be the bank payment method")
-
-    def test_fast_payment_validation_from_product_screen_with_automatic_receipt_printing(self):
-        pos_printer = self.env['pos.printer'].create({
-            'name': 'Printer',
-            'printer_type': 'epson_epos',
-            'printer_ip': '0.0.0.0',
-            'use_type': 'receipt',
-        })
-        self.main_pos_config.write({
-            'use_fast_payment': True,
-            'fast_payment_method_ids': [(6, 0, self.bank_payment_method.ids)],
-            'iface_print_auto': True,
-            'other_devices': True,
-            'receipt_printer_ids': [Command.set(pos_printer.ids)],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_fast_payment_validation_from_product_screen_with_automatic_receipt_printing')
-        order1 = self.main_pos_config.current_session_id.order_ids[0]
-        order2 = self.main_pos_config.current_session_id.order_ids[1]
-        self.assertEqual(order1.state, 'paid', "The order should be paid after the fast payment validation")
-        self.assertEqual(len(order1.payment_ids), 1, "There should be one payment line used for the fast payment")
-        self.assertEqual(order1.payment_ids.payment_method_id, self.bank_payment_method, "The payment method used should be the bank payment method")
-        self.assertEqual(order2.state, 'paid', "The order should be paid")
-        self.assertEqual(len(order2.payment_ids), 1, "There should be one payment line")
-        self.assertEqual(order2.payment_ids.payment_method_id, self.bank_payment_method, "The payment method used should be the bank payment method")
-
-    def test_consistent_refund_process_between_frontend_and_backend(self):
-        """
-        Ensure that the partial refund process is consistent between the frontend and backend.
-        This includes validating the refund order creation, amount, state, and payment processing.
-        """
-        # Open POS UI with the POS user
-        pricelists = self.env['product.pricelist'].create([
-            {'name': 'Test Pricelist'},
-            {'name': 'Percentage Pricelist'},
-        ])
-        self.main_pos_config.write({
-            'available_pricelist_ids': [Command.set(pricelists.ids)],
-            'pricelist_id': pricelists[0].id,
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-
-        # Run the POS tour simulating a partial refund
-        self.start_pos_tour('test_consistent_refund_process_between_frontend_and_backend')
-
-        # Fetch orders created in the current POS session
-        orders = self.env['pos.order'].search([
-            ('session_id', '=', self.main_pos_config.current_session_id.id)
-        ])
-        self.assertEqual(len(orders), 2, "Expected two orders: original and refund.")
-        original_order = next(o for o in orders if o.amount_total > 0)
-        frontend_refund_order = next(o for o in orders if o.amount_total < 0)
-        self.assertEqual(
-            frontend_refund_order.pricelist_id.id,
-            original_order.pricelist_id.id,
-            "Refund order pricelist should be the original order's pricelist."
-        )
-
-        # Perform refund on order and retrieve the resulting draft refund order
-        refund_action = original_order.refund()
-        refund_order = self.env['pos.order'].browse(refund_action['res_id'])
-
-        # Validate the refund order is in draft and has correct negative total
-        self.assertEqual(refund_order.state, 'draft', "Refund order should be in draft state.")
-        self.assertEqual(refund_order.amount_total, -4, "Refund order total should be -4.")
-
-        # Create a payment for the refund using the configured bank method
-        payment_context = {
-            "active_ids": refund_order.ids,
-            "active_id": refund_order.id
-        }
-        refund_payment = self.env['pos.make.payment'].with_context(**payment_context).create({
-            'amount': refund_order.amount_total,
-            'payment_method_id': self.bank_payment_method.id,
-        })
-
-        # Validate and finalize the refund payment
-        refund_payment.with_context(**payment_context).check()
-        self.assertEqual(refund_order.state, 'paid', "Refund order should be marked as paid.")
-
-    def test_paid_order_with_archived_product_loads(self):
-        """ Test that a paid order with archived products can be loaded in the POS. """
-
-        archived_product = self.env['product.product'].create({
-            'name': 'Archived Product',
-            'available_in_pos': True,
-            'list_price': 10.0,
-            'taxes_id': False,
-            'active': False,  # Archived product
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.env['pos.order'].create({
-            'config_id': self.main_pos_config.id,
-            'session_id': self.main_pos_config.current_session_id.id,
-            'company_id': self.main_pos_config.company_id.id,
-            'amount_total': 10.0,
-            'amount_paid': 10.0,
-            'amount_tax': 0.0,
-            'amount_return': 0.0,
-            'to_invoice': False,
-            'partner_id': False,
-            'pricelist_id': self.main_pos_config.pricelist_id.id,
-            'pos_reference': '1000-004-00002',
-            'name': 'Order 0002',
-            'state': 'paid',
-            'lines': [(0, 0, {
-                'name': 'Line 0001',
-                'product_id': archived_product.id,
-                'price_unit': 10.00,
-                'discount': 0,
-                'qty': 1,
-                'tax_ids': False,
-                'price_subtotal': 10.00,
-                'price_subtotal_incl': 10.00,
-            })],
-        })
-
-        self.start_tour(f"/pos/ui?config_id={self.main_pos_config.id}", 'test_paid_order_with_archived_product_loads', login="pos_user")
-
-    def test_delete_line(self):
-        """ Test that deleting a line in the POS through the popup works correctly. """
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_delete_line')
-
-    def test_order_invoice_search(self):
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.pos_user.group_ids = [Command.link(self.env.ref('account.group_account_invoice').id)]
-        self.start_tour("/pos/ui/%d" % self.main_pos_config.id, 'test_order_invoice_search', login="pos_user")
-
-    def test_automatic_receipt_printing(self):
-        self.main_pos_config.write({
-            'iface_print_auto': True,
-            'receipt_printer_ids': [self.printer.id],
-            'other_devices': True,
-        })
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_automatic_receipt_printing', login="pos_user")
-
-    def test_load_pos_demo_data(self):
-        """ Test that the demo data can be loaded by admin but not by user. """
-
-        if loaded_demo_data(self.env):
-            self.skipTest('Cannot test with demo data.')
-
-        # archive existing product records
-        archive_products(self.env)
-
-        # cannot load by pos user
-        self.start_pos_tour('test_load_pos_demo_data_by_pos_user', login='pos_user')
-        products = self.env['product.template'].search_count([('available_in_pos', '=', True)])
-        self.assertFalse(products, 'Demo data should not be loaded by user.')
-
-        # Member role with POS Administrator access
-        self.pos_user.write({'group_ids': [
-            Command.set(
-                [
-                    self.env.ref('base.group_user').id,
-                    self.env.ref('point_of_sale.group_pos_manager').id,
-                    self.env.ref('account.group_account_manager').id,
-                ]
-            )
-        ]})
-        self.start_pos_tour('test_load_pos_demo_data_with_member_role', login='pos_user')
-        products = self.env['product.template'].search_count([('available_in_pos', '=', True)])
-        self.assertFalse(products, 'Demo data should not be loaded by user with member role.')
-
-    def test_combo_variant_mix(self):
-        color_attribute = self.env['product.attribute'].create({
-            'name': 'Color',
-            'value_ids': [
-                Command.create({'name': 'Red'}),
-                Command.create({'name': 'Blue'})
-            ],
-            'create_variant': 'no_variant',
-        })
-        size_attribute = self.env['product.attribute'].create({
-            'name': 'Size',
-            'value_ids': [
-                Command.create({'name': 'Small'}),
-                Command.create({'name': 'Large'})
-            ],
-            'create_variant': 'always',
-        })
-
-        product_template = self.env['product.template'].create({
-            'name': 'Test Product',
-            'available_in_pos': True,
-            'list_price': 10,
-            'taxes_id': False,
-            'attribute_line_ids': [
-                Command.create({
-                    'attribute_id': color_attribute.id,
-                    'value_ids': [Command.link(id) for id in color_attribute.value_ids.ids]
-                }),
-                Command.create({
-                    'attribute_id': size_attribute.id,
-                    'value_ids': [Command.link(id) for id in size_attribute.value_ids.ids]
-                })
-            ]
-        })
-
-        combo = self.env['product.combo'].create({
-            'name': 'Test Combo',
-            'combo_item_ids': [
-                Command.create({
-                    'product_id': product_template.product_variant_ids[0].id,
-                    'extra_price': 0,
-                }),
-                Command.create({
-                    'product_id': product_template.product_variant_ids[1].id,
-                    'extra_price': 0,
-                }),
-            ]
-        })
-        self.env['product.template'].create({
-            'name': 'Test Product Combo',
-            'available_in_pos': True,
-            'list_price': 20,
-            'taxes_id': False,
-            'type': 'combo',
-            'combo_ids': [Command.link(combo.id)],
-        })
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_combo_variant_mix', login="pos_user")
-
-    def test_cross_exclusion_attribute_values(self):
-        """ If you create a product with two attributes and 2 values for each attribute, and you exclude one value of the first attribute with one value of the second attribute
-        and vice versa, you should still be able to select the other values of the attributes. """
-        self.attribute_1 = self.env['product.attribute'].create({
-            'name': 'attribute_1',
-            'create_variant': 'no_variant',
-        })
-
-        self.attribute_2 = self.env['product.attribute'].create({
-            'name': 'attribute_2',
-            'create_variant': 'no_variant',
-        })
-
-        self.attribute_1_value_1 = self.env['product.attribute.value'].create({
-            'name': 'attribute_1_value_1',
-            'attribute_id': self.attribute_1.id,
-        })
-        self.attribute_1_value_2 = self.env['product.attribute.value'].create({
-            'name': 'attribute_1_value_2',
-            'attribute_id': self.attribute_1.id,
-        })
-        self.attribute_2_value_1 = self.env['product.attribute.value'].create({
-            'name': 'attribute_2_value_1',
-            'attribute_id': self.attribute_2.id,
-        })
-        self.attribute_2_value_2 = self.env['product.attribute.value'].create({
-            'name': 'attribute_2_value_2',
-            'attribute_id': self.attribute_2.id,
-        })
-
-        self.test_product_1 = self.env['product.template'].create({
-            'name': 'Test Product 1',
-            'available_in_pos': True,
-            'list_price': 10.0,
-            'attribute_line_ids': [
-                (0, 0, {
-                    'attribute_id': self.attribute_1.id,
-                    'value_ids': [(6, 0, [self.attribute_1_value_1.id, self.attribute_1_value_2.id])],
-                }),
-                (0, 0, {
-                    'attribute_id': self.attribute_2.id,
-                    'value_ids': [(6, 0, [self.attribute_2_value_1.id, self.attribute_2_value_2.id])],
-                }),
-            ],
-        })
-
-        # Test the exclusion of attribute values
-        ptav_1_1 = self.test_product_1.attribute_line_ids.filtered(lambda l: l.attribute_id.id == self.attribute_1.id).product_template_value_ids.filtered(lambda v: v.product_attribute_value_id.id == self.attribute_1_value_1.id)
-        ptav_1_2 = self.test_product_1.attribute_line_ids.filtered(lambda l: l.attribute_id.id == self.attribute_1.id).product_template_value_ids.filtered(lambda v: v.product_attribute_value_id.id == self.attribute_1_value_2.id)
-        ptav_2_2 = self.test_product_1.attribute_line_ids.filtered(lambda l: l.attribute_id.id == self.attribute_2.id).product_template_value_ids.filtered(lambda v: v.product_attribute_value_id.id == self.attribute_2_value_2.id)
-        ptav_2_1 = self.test_product_1.attribute_line_ids.filtered(lambda l: l.attribute_id.id == self.attribute_2.id).product_template_value_ids.filtered(lambda v: v.product_attribute_value_id.id == self.attribute_2_value_1.id)
-
-        ptav_1_1.excluded_value_ids = [Command.link(ptav_2_1.id)]
-        ptav_1_2.excluded_value_ids = [Command.link(ptav_2_2.id)]
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_cross_exclusion_attribute_values')
-
-    def test_custom_attribute_alone_displayed(self):
-        """
-        Tests that if product configurator will be shown if any of the
-        attributes have a free text field, even if there is only one
-        possible selection for every attributes.
-        """
-        attribute_custom = self.env['product.attribute'].create({
-            'name': 'Custom',
-            'display_type': 'radio',
-            'create_variant': 'no_variant',
-        })
-        attribute_value_custom = self.env['product.attribute.value'].create({
-            'name': 'Custom',
-            'attribute_id': attribute_custom.id,
-            'is_custom': True,
-        })
-        self.test_product_1 = self.env['product.template'].create({
-            'name': 'Only Custom',
-            'available_in_pos': True,
-            'list_price': 10.0,
-            'attribute_line_ids': [
-                Command.create({
-                    'attribute_id': attribute_custom.id,
-                    'value_ids': [Command.set([attribute_value_custom.id])],
-                }),
-            ],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_custom_attribute_alone_displayed')
-
-    def test_weight_product(self):
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_weight_product')
-        order = self.env['pos.order'].search([], limit=1)
-        self.assertEqual(len(order.lines), 2, "There should be two order lines")
-        self.assertEqual(order.lines[0].price_subtotal_incl, 40, "The price unit should be 40")
-        self.assertEqual(order.lines[0].qty, 4, "The quantity should be 4")
-        self.assertEqual(order.lines[1].price_subtotal_incl, 40, "The price unit should be 40")
-        self.assertEqual(order.lines[1].qty, 1, "The quantity should be 1")
-
-    def test_preset_customer_selection(self):
-        self.preset_delivery = self.env['pos.preset'].create({
-            'name': 'Delivery',
-            'identification': 'address',
-        })
-        self.main_pos_config.write({
-            'use_presets': True,
-            'default_preset_id': self.preset_delivery.id,
-            'available_preset_ids': [(6, 0, [self.preset_delivery.id])],
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_preset_customer_selection')
-
-    def test_pos_large_amount_confirmation_dialog(self):
-        """Test that the Large amount confirmation dialog appears
-        and closes properly after clicking 'OK'."""
-        self.env['product.product'].create({
-            'name': 'Overpay Test Product',
-            'list_price': 1.0,
-            'available_in_pos': True,
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_pos_large_amount_confirmation_dialog')
-
-    def test_add_money_button_with_different_decimal_separator(self):
-        """
-        Tests that the buttons such as +10 or +50 work even in languages such as
-        french that have ',' as a decimal separator.
-        """
-        lang = self.env['res.lang'].search([('code', '=', self.pos_user.lang)])
-        lang.write({'thousands_sep': '.', 'decimal_point': ','})
-        self.start_tour(f"/pos/ui?config_id={self.main_pos_config.id}", 'test_add_money_button_with_different_decimal_separator', login="pos_user")
-
-    def test_convert_orderlines_to_combo(self):
-        setup_product_combo_items(self)
-
-        pos_category_ids = self.desks_combo.combo_item_ids[0].product_id.pos_categ_ids.ids
-
-        # Create a second combo
-        combo_product_1 = self.env["product.product"].create(
-            {
-                "name": "Second Product 1",
-                "is_storable": True,
-                "available_in_pos": True,
-                "list_price": 10,
-                "taxes_id": [],
-                "pos_categ_ids": [(6, 0, pos_category_ids)],
-            }
-        )
-
-        combo_product_2 = self.env["product.product"].create(
-            {
-                "name": "Second Product 2",
-                "is_storable": True,
-                "available_in_pos": True,
-                "list_price": 11,
-                "taxes_id": [],
-                "pos_categ_ids": [(6, 0, pos_category_ids)],
-            }
-        )
-
-        combo_product_3 = self.env["product.product"].create(
-            {
-                "name": "Second Product 3",
-                "is_storable": True,
-                "available_in_pos": True,
-                "list_price": 16,
-                "taxes_id": [],
-                "pos_categ_ids": [(6, 0, pos_category_ids)],
-            }
-        )
-
-        first_combo = self.env["product.combo"].create(
-            {
-                "name": "First Combo",
-                "combo_item_ids": [
-                    Command.create({
-                        "product_id": combo_product_1.id,
-                        "extra_price": 0,
-                    }),
-                    Command.create({
-                        "product_id": combo_product_2.id,
-                        "extra_price": 0,
-                    }),
-                    Command.create({
-                        "product_id": combo_product_3.id,
-                        "extra_price": 2,
-                    }),
-                ],
-            }
-        )
-
-        combo_product_4 = self.env["product.product"].create(
-            {
-                "name": "Second Product 4",
-                "is_storable": True,
-                "available_in_pos": True,
-                "list_price": 20,
-                "taxes_id": [],
-                "pos_categ_ids": [(6, 0, pos_category_ids)],
-            }
-        )
-
-        combo_product_5 = self.env["product.product"].create(
-            {
-                "name": "Second Product 5",
-                "is_storable": True,
-                "available_in_pos": True,
-                "list_price": 25,
-                "taxes_id": [],
-                "pos_categ_ids": [(6, 0, pos_category_ids)],
-            }
-        )
-
-        second_combo = self.env["product.combo"].create(
-            {
-                "name": "Second Combo",
-                "combo_item_ids": [
-                    Command.create({
-                        "product_id": combo_product_4.id,
-                        "extra_price": 0,
-                    }),
-                    Command.create({
-                        "product_id": combo_product_5.id,
-                        "extra_price": 2,
-                    }),
-                ],
-            }
-        )
-
-        combo_product_6 = self.env["product.product"].create(
-            {
-                "name": "Second Product 6",
-                "is_storable": True,
-                "available_in_pos": True,
-                "list_price": 30,
-                "taxes_id": [],
-                "pos_categ_ids": [(6, 0, pos_category_ids)],
-            }
-        )
-
-        combo_product_7 = self.env["product.product"].create(
-            {
-                "name": "Second Product 7",
-                "is_storable": True,
-                "available_in_pos": True,
-                "list_price": 32,
-                "taxes_id": [],
-                "pos_categ_ids": [(6, 0, pos_category_ids)],
-            }
-        )
-
-        combo_product_8 = self.env["product.product"].create(
-            {
-                "name": "Second Product 8",
-                "is_storable": True,
-                "available_in_pos": True,
-                "list_price": 40,
-                "taxes_id": [],
-                "pos_categ_ids": [(6, 0, pos_category_ids)],
-            }
-        )
-
-        combo_product_9 = self.env["product.product"].create(
-            {
-                "name": "Second Product 9",
-                "is_storable": True,
-                "available_in_pos": True,
-                "list_price": 50,
-                "taxes_id": [],
-                "pos_categ_ids": [(6, 0, pos_category_ids)],
-            }
-        )
-
-        chair_color_attribute = self.env['product.attribute'].create({
-            'name': 'Color',
-            'display_type': 'color',
-            'create_variant': 'no_variant',
-        })
-        chair_color_red = self.env['product.attribute.value'].create({
-            'name': 'Red',
-            'attribute_id': chair_color_attribute.id,
-            'html_color': '#ff0000',
-        })
-        chair_color_blue = self.env['product.attribute.value'].create({
-            'name': 'Blue',
-            'attribute_id': chair_color_attribute.id,
-            'html_color': '#0000ff',
-        })
-        self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': combo_product_9.product_tmpl_id.id,
-            'attribute_id': chair_color_attribute.id,
-            'value_ids': [(6, 0, [chair_color_red.id, chair_color_blue.id])]
-        })
-
-        third_combo = self.env["product.combo"].create(
-            {
-                "name": "Third Combo",
-                "combo_item_ids": [
-                    Command.create({
-                        "product_id": combo_product_6.id,
-                        "extra_price": 0,
-                    }),
-                    Command.create({
-                        "product_id": combo_product_7.id,
-                        "extra_price": 0,
-                    }),
-                    Command.create({
-                        "product_id": combo_product_8.id,
-                        "extra_price": 5,
-                    }),
-                    Command.create({
-                        "product_id": combo_product_9.id,
-                        "extra_price": 0,
-                    }),
-                ],
-            }
-        )
-
-        third_combo_with_upsell = self.env["product.combo"].create(
-            {
-                "name": "Third Combo",
-                "is_upsell": True,
-                "qty_free": 0,
-                "combo_item_ids": [
-                    Command.create({
-                        "product_id": combo_product_6.id,
-                        "extra_price": 0,
-                    }),
-                    Command.create({
-                        "product_id": combo_product_7.id,
-                        "extra_price": 0,
-                    }),
-                    Command.create({
-                        "product_id": combo_product_8.id,
-                        "extra_price": 5,
-                    }),
-                    Command.create({
-                        "product_id": combo_product_9.id,
-                        "extra_price": 0,
-                    }),
-                ],
-            }
-        )
-
-        # Create Office Combo
-        self.office_combo = self.env["product.product"].create(
-            {
-                "available_in_pos": True,
-                "list_price": 50,
-                "name": "Second Combo Product",
-                "type": "combo",
-                "uom_id": self.env.ref("uom.product_uom_unit").id,
-                "combo_ids": [
-                    (6, 0, [first_combo.id, second_combo.id, third_combo.id])
-                ],
-                "pos_categ_ids": [(6, 0, pos_category_ids)],
-            }
-        )
-
-        self.third_combo_product = self.env["product.product"].create(
-            {
-                "available_in_pos": True,
-                "list_price": 50,
-                "name": "Third Combo Product",
-                "type": "combo",
-                "uom_id": self.env.ref("uom.product_uom_unit").id,
-                "combo_ids": [
-                    (6, 0, [first_combo.id, second_combo.id, third_combo_with_upsell.id])
-                ],
-                "pos_categ_ids": [(6, 0, pos_category_ids)],
-            }
-        )
-
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_convert_orderlines_to_combo', login="pos_user")
-        self.start_pos_tour('test_convert_orderlines_to_combo_with_upsell', login="pos_user")
-
-    def test_sync_from_ui_one_by_one(self):
-        """
-        Sync from UI is now syncing orders one by one.
-        sync_from_ui should be called 6 times in this tour (6 orders created).
-        """
-
-        pos_order = self.env.registry.models['pos.order']
-        sync_counter = {'count': 0}
-
-        @api.model
-        def sync_from_ui_patch(self, orders):
-            sync_counter['count'] += 1
-            return super(pos_order, self).sync_from_ui(orders)
-
-        with patch.object(pos_order, "sync_from_ui", sync_from_ui_patch):
-            self.start_pos_tour("test_sync_from_ui_one_by_one", login="pos_user")
-            self.assertEqual(sync_counter['count'], 6)
-
-    def test_refund_line_keep_attributes(self):
-        """
-        Tests that when refunding an order that has lines that are variants, the new line keeps
-        this variant and displays it.
-        """
-        product_test = self.env['product.product'].create({
-            'name': 'Donut',
-            'list_price': 10,
-            'available_in_pos': True,
-            'taxes_id': False,
-        })
-        attribute = self.env['product.attribute'].create({
-            'name': 'Flavor',
-            'create_variant': 'always',
-        })
-        attribute_value_1 = self.env['product.attribute.value'].create({
-            'name': 'Sugar',
-            'attribute_id': attribute.id,
-        })
-        attribute_value_2 = self.env['product.attribute.value'].create({
-            'name': 'Chocolate',
-            'attribute_id': attribute.id,
-        })
-        self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product_test.product_tmpl_id.id,
-            'attribute_id': attribute.id,
-            'value_ids': [(6, 0, [attribute_value_1.id, attribute_value_2.id])],
-        })
-        self.start_pos_tour("test_refund_line_keep_attributes")
-
-    def test_pos_snooze(self):
-        """Test that the available switch is on for available products
-        and off for snoozed products. """
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_pos_snooze')
-
-    def test_set_opening_note_without_cash_method(self):
-        cash_method = self.main_pos_config.payment_method_ids.filtered(lambda pm: pm.is_cash_count)
-        self.main_pos_config.payment_method_ids -= cash_method
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        current_session = self.main_pos_config.current_session_id
-        self.start_pos_tour('test_set_opening_note_without_cash_method')
-        self.assertEqual(current_session.opening_notes, 'Opening Notes')
-
-    def test_orderline_merge_with_higher_price_precision(self):
-        """ Test that orderline merging works correctly when product price has a higher precision than the currency. """
-        self.env['decimal.precision'].search([('name', '=', 'Product Price')]).digits = 3
-
-        self.env['product.product'].create({
-            'name': 'High Precision Product',
-            'list_price': 8.245,
-            'taxes_id': False,
-            'available_in_pos': True,
-        })
-
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_orderline_merge_with_higher_price_precision', login="pos_user")
-
-    def test_product_configurator_price(self):
-        """ Test that the product configurator displays the correct price when selecting attributes that impact the price. """
-        self.env['product.template'].search([('available_in_pos', '=', True)]).active = False
-        tax_10 = self.env['account.tax'].create({
-            'name': 'Tax 10%',
-            'amount': 10,
-        })
-        fiscal_position = self.env['account.fiscal.position'].create({
-            'name': 'Include to Exclude',
-        })
-        tax_10 = self.env['account.tax'].create({
-            'name': 'Tax 10 Excluded',
-            'amount': 10,
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-            'price_include_override': 'tax_excluded',
-        })
-        self.env['account.tax'].create({
-            'name': 'Tax 10 Included',
-            'amount': 10,
-            'amount_type': 'percent',
-            'type_tax_use': 'sale',
-            'price_include_override': 'tax_included',
-            'fiscal_position_ids': fiscal_position,
-            'original_tax_ids': tax_10,
-        })
-        product = self.env['product.template'].create({
-            'name': 'Configurable Product',
-            'available_in_pos': True,
-            'list_price': 10.0,
-            'taxes_id': [(6, 0, [tax_10.id])],
-        })
-        size_attribute = self.env['product.attribute'].create({
-            'name': 'Size',
-            'create_variant': 'always',
-        })
-        color_attribute = self.env['product.attribute'].create({
-            'name': 'Color',
-            'create_variant': 'no_variant',
-        })
-        small_size_value, large_size_value = self.env['product.attribute.value'].create([{
-            'name': 'Small',
-            'attribute_id': size_attribute.id,
-        }, {
-            'name': 'Large',
-            'attribute_id': size_attribute.id,
-        }])
-        red_color_value, blue_color_value = self.env['product.attribute.value'].create([{
-            'name': 'Red',
-            'attribute_id': color_attribute.id,
-        }, {
-            'name': 'Blue',
-            'attribute_id': color_attribute.id,
-        }])
-        size_line = self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product.id,
-            'attribute_id': size_attribute.id,
-            'value_ids': [(6, 0, [small_size_value.id, large_size_value.id])],
-        })
-        size_line.product_template_value_ids[1].price_extra = 1
-        color_line = self.env['product.template.attribute.line'].create({
-            'product_tmpl_id': product.id,
-            'attribute_id': color_attribute.id,
-            'value_ids': [(6, 0, [red_color_value.id, blue_color_value.id])],
-        })
-        color_line.product_template_value_ids[0].price_extra = 2
-        color_line.product_template_value_ids[1].price_extra = 3
-
-        pricelist_1, pricelist_2 = self.env['product.pricelist'].create([{
-            'name': 'Pricelist 1',
-        }, {
-            'name': 'Pricelist 2',
-            'item_ids': [(0, 0, {
-                'applied_on': '1_product',
-                'product_tmpl_id': product.id,
-                'fixed_price': 20.0,
-            })],
-        }])
-        self.main_pos_config.write({
-            'available_pricelist_ids': [(6, 0, [pricelist_1.id, pricelist_2.id])],
-            'pricelist_id': pricelist_1.id,
-            'tax_regime_selection': True,
-            'fiscal_position_ids': [(6, 0, [fiscal_position.id])],
-        })
-        self.start_tour("/pos/ui?config_id=%d" % self.main_pos_config.id, 'test_product_configurator_price', login="pos_user")
-
-    def test_combo_no_free_item(self):
-        """ Test a product combo with no free item allowed. """
-        setup_product_combo_items(self)
-        for combo_item in self.office_combo.combo_ids:
-            combo_item.write({
-                'is_upsell': True,
-                'qty_free': 0,
-                'qty_max': 5,
-            })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_combo_no_free_item')
-
-    def test_not_available_pricelist_not_set_on_order(self):
-        """ Test that when the pricelist is not available, it is not set on the order """
-        not_available_pricelist, available_pricelist = self.env['product.pricelist'].create([{
-            'name': 'Not Available Pricelist',
-        }, {
-            'name': 'Available Pricelist',
-        }])
-
-        self.main_pos_config.write({
-            'available_pricelist_ids': [(4, available_pricelist.id)],
-            'pricelist_id': available_pricelist.id,
-        })
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        pos_session = self.main_pos_config.current_session_id
-
-        partner = self.env['res.partner'].create({
-            'name': 'AA Customer',
-            'property_product_pricelist': not_available_pricelist.id,
-        })
-
-        order = self.env['pos.order'].create({
-            'company_id': self.env.company.id,
-            'session_id': pos_session.id,
-            'partner_id': partner.id,
-            'config_id': self.main_pos_config.id,
-            'lines': [(0, 0, {
-                'name': 'OL/0001',
-                'product_id': self.wall_shelf.product_variant_ids[0].id,
-                'price_unit': 10.00,
-                'discount': 0,
-                'qty': 1,
-                'tax_ids': False,
-                'price_subtotal': 10.00,
-                'price_subtotal_incl': 10.00,
-            })],
-            'pricelist_id': not_available_pricelist.id,
-            'amount_paid': 10.00,
-            'amount_total': 10.00,
-            'amount_tax': 0.0,
-            'amount_return': 0.0,
-            'to_invoice': False,
-            'pos_reference': 'Test/0001',
-        })
-        order.action_pos_order_paid()
-
-        self.start_tour(f"/pos/ui?config_id={self.main_pos_config.id}", 'test_not_available_pricelist_not_set_on_order', login="pos_user")
-
-        created_order = self.env['pos.order'].search([('partner_id', '=', partner.id)], limit=1)
-        self.assertNotEqual(created_order.pricelist_id, not_available_pricelist)
-
-    def test_pos_open_ui_button(self):
-        """ Test the Open Register button click behavior in the dashboard. """
-        self.start_tour("/odoo/point-of-sale", 'test_pos_open_ui_button', login="pos_user")
-
-    def test_customer_search_prefilled_on_create(self):
-        self.main_pos_config.with_user(self.pos_user).open_ui()
-        self.start_pos_tour('test_customer_search_prefilled_on_create')
 
 
 # This class just runs the same tests as above but with mobile emulation
